Removed old cross-section code
