+++ ubcc1pi_standalone/Helpers/CrossSectionHelper.cxx
@@ -6,1494 +6,112 @@
 
 #include "ubcc1pi_standalone/Helpers/CrossSectionHelper.h"
 
+#include "ubcc1pi_standalone/Helpers/AnalysisHelper.h"
+
 #include <stdexcept>
 
 namespace ubcc1pi
 {
 
-CrossSectionHelper::CrossSection::CrossSection(const std::vector<float> &binEdges, const bool hasUnderflow, const bool hasOverflow, const bool scaleByBinWidth, const InputData &inputData) :
-    m_binEdges(binEdges),
-    m_hasUnderflow(hasUnderflow),
-    m_hasOverflow(hasOverflow),
-    m_scaleByBinWidth(scaleByBinWidth),
-    m_inputData(inputData),
-    m_allEventsNuEnergyNomTrue(this->GetEmptyFluxHist()),
-    m_dataSelectedReco(this->GetEmptyHist1D()),
-    m_signalSelectedNomRecoTrue(this->GetEmptyHist2D()),
-    m_signalAllNomTrue(this->GetEmptyHist1D()),
-    m_backgroundSelectedNomReco(this->GetEmptyHist1D()),
-    m_backgroundSelectedNonOverlayDetVarReco(this->GetEmptyHist1D()),
-    m_shouldResetCache(true)
-{
-    // Setup the histograms for the universe variations
-    for (const auto &[name, nUniverses] : m_inputData.m_systUniverseSizesMap)
-    {
-        auto &allEventsNuEnergyTrueVect = m_allEventsNuEnergyTrue[name];
-        auto &signalSelectedRecoTrueVect = m_signalSelectedRecoTrue[name];
-        auto &signalAllTrueVect = m_signalAllTrue[name];
-        auto &backgroundSelectedRecoVect = m_backgroundSelectedReco[name];
+CrossSectionHelper::FluxReweightor::FluxReweightor(const std::vector<float> &binEdges, const std::vector<float> &binValuesNominal, const SystDimensionsMap &fluxWeightsDimensions) :
+    m_dimensions(fluxWeightsDimensions),
+    m_pFluxNominal(CrossSectionHelper::GetTH1F(binEdges)),
+    m_pSpectrumNominal(CrossSectionHelper::GetTH1F(binEdges)),
+    m_spectrumVariations(CrossSectionHelper::GetSystTH1FMap(binEdges, fluxWeightsDimensions)),
 
-        for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
+    // Define the cache-function that gets the integrated flux variations
+    // ATTN it might look a bit odd to some to define a function in the constructor! To be clear, here we are initalizing the member
+    // variable named m_getIntegratedFluxVariationFunction. This is a SystCacheFunction that takes a lambda function as it's first paramter
+    // and stores the logic for later. When the user calls GetIntegratedFluxVariation() we ask the SystCacheFunction to check if it has the
+    // value we are looking for in it's cache. If the answer is yes, then we just return the cached value. If the answer is no, then we call
+    // the logic (defined in the lambda function below), store the result in the cache for later and return it.
+    m_getIntegratedFluxVariationFunction([&](const std::string &paramName, const unsigned int universeIndex)
+    {
+        // Setup a new histogram to hold the reweighted flux
+        auto pReweightedFlux = CrossSectionHelper::GetTH1F(binEdges);
+
+        // Get the event rate spectrum in this universe
+        const auto pSpectrumVariation = m_spectrumVariations.at(paramName).at(universeIndex);
+
+        // Loop over each bin
+        const auto nBins = binEdges.size() - 1;
+        for (unsigned iBin = 1; iBin <= nBins; ++iBin)
         {
-            allEventsNuEnergyTrueVect.push_back(this->GetEmptyFluxHist());
-            signalSelectedRecoTrueVect.push_back(this->GetEmptyHist2D());
-            signalAllTrueVect.push_back(this->GetEmptyHist1D());
-            backgroundSelectedRecoVect.push_back(this->GetEmptyHist1D());
-        }
-    }
+            // Get bin value for the nominal spectrum and the specturm in this universe
+            const auto nEventsNominal = m_pSpectrumNominal->GetBinContent(iBin);
+            const auto nEventsVariation = pSpectrumVariation->GetBinContent(iBin);
 
-    // Check that each run ID of the detector variations has a CV samples
-    std::map<std::string, bool> runIdToHasCVMap;
-    for (const auto &[runId, detVarParam] : m_inputData.m_detVarParameters)
-    {
-        // Add this run ID to the map if we haven't seen it before
-        if (runIdToHasCVMap.find(runId) == runIdToHasCVMap.end())
-            runIdToHasCVMap.emplace(runId, false);
+            // Get the ratio of the variation to the nominal (if not possible use a unit weight)
+            const auto weight = (nEventsNominal <= std::numeric_limits<float>::epsilon() ? 1.f : (nEventsVariation / nEventsNominal) );
 
-        runIdToHasCVMap.at(runId) = (runIdToHasCVMap.at(runId) || detVarParam == m_cvString);
-    }
+            // Apply this weight to the nominal flux
+            const auto nominalFlux = m_pFluxNominal->GetBinContent(iBin);
+            const auto reweightedFlux = nominalFlux * weight;
 
-    for (const auto &[runId, hasCV] : runIdToHasCVMap)
-    {
-        if (!hasCV)
-            throw std::invalid_argument("CrossSection::CrossSection - Run: " + runId + " doesn't have a central value. Expected to see a sample with name: " + m_cvString);
-    }
-
-    // Setup the histograms for the detector variations
-    for (const auto &[runId, detVarParam] : m_inputData.m_detVarParameters)
-    {
-        m_signalSelectedDetVarRecoTrue[runId][detVarParam] = this->GetEmptyHist2D();
-        m_signalAllDetVarTrue[runId][detVarParam] = this->GetEmptyHist1D();
-        m_backgroundSelectedOverlayDetVarReco[runId][detVarParam] = this->GetEmptyHist1D();
-    }
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-bool CrossSectionHelper::CrossSection::HasUnderflowBin() const
-{
-    return m_hasUnderflow;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-bool CrossSectionHelper::CrossSection::HasOverflowBin() const
-{
-    return m_hasOverflow;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::CrossSection::AddNeutrinoEvent(const float trueNuEnergy, const float nominalWeight, const SystematicWeightsMap &systWeightsMap)
-{
-    // Check the validity of the input map if systematic weights
-    if (!CrossSectionHelper::CheckSystematicWeightsMapDimensions(systWeightsMap, m_inputData.m_systUniverseSizesMap))
-        throw std::invalid_argument("CrossSection::AddNeutrinoEvent - Input systematic weights map has the wrong dimensions!");
-
-    // Add this event in the nominal universe
-    m_allEventsNuEnergyNomTrue->Fill(trueNuEnergy, nominalWeight);
-
-    // Add the event in each systematic universe
-    for (const auto &[name, nUniverses] : m_inputData.m_systUniverseSizesMap)
-    {
-        auto &allEventsNuEnergyTrueVect = m_allEventsNuEnergyTrue.at(name);
-
-        const auto &weights = systWeightsMap.at(name);
-        for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
-        {
-            const auto universeWeight = weights.at(iUni);
-            allEventsNuEnergyTrueVect.at(iUni)->Fill(trueNuEnergy, nominalWeight * universeWeight);
-        }
-    }
-
-    // Mark that the inputs have changed and any cached values need to be re-calculated
-    m_shouldResetCache = true;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::CrossSection::AddSignalEvent(const float trueValue, const float recoValue, const bool isSelected, const float nominalWeight, const SystematicWeightsMap &systWeightsMap)
-{
-    // Check the validity of the input map if systematic weights
-    if (!CrossSectionHelper::CheckSystematicWeightsMapDimensions(systWeightsMap, m_inputData.m_systUniverseSizesMap))
-        throw std::invalid_argument("CrossSection::AddSignalEvent - Input systematic weights map has the wrong dimensions!");
-
-    // Add this event in the nominal universe
-    m_signalAllNomTrue->Fill(trueValue, nominalWeight);
-
-    if (isSelected)
-        m_signalSelectedNomRecoTrue->Fill(trueValue, recoValue, nominalWeight);
-
-    // Add the event in each systematic universe
-    for (const auto &[name, nUniverses] : m_inputData.m_systUniverseSizesMap)
-    {
-        auto &signalAllTrueVect = m_signalAllTrue.at(name);
-        auto &signalSelectedRecoTrueVect = m_signalSelectedRecoTrue.at(name);
-
-        const auto &weights = systWeightsMap.at(name);
-        for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
-        {
-            const auto universeWeight = weights.at(iUni);
-            signalAllTrueVect.at(iUni)->Fill(trueValue, nominalWeight * universeWeight);
-
-            if (isSelected)
-                signalSelectedRecoTrueVect.at(iUni)->Fill(trueValue, recoValue, nominalWeight * universeWeight);
-        }
-    }
-
-    // Mark that the inputs have changed and any cached values need to be re-calculated
-    m_shouldResetCache = true;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::CrossSection::AddSignalEventDetVar(const float trueValue, const float recoValue, const bool isSelected, const float nominalWeight, const std::string &runId, const std::string &detectorVariationParameter)
-{
-    // Check if this is a value run-ID / detector variation parameter
-    if (std::find(m_inputData.m_detVarParameters.begin(), m_inputData.m_detVarParameters.end(), std::pair<std::string, std::string> (runId, detectorVariationParameter)) == m_inputData.m_detVarParameters.end())
-    {
-        throw std::invalid_argument("CrossSection::AddSignalEventDetVar - Unknown detector variation parameter: " + detectorVariationParameter + " for run: " + runId);
-    }
-
-    // Add this event to the counter for all signal events
-    m_signalAllDetVarTrue.at(runId).at(detectorVariationParameter)->Fill(trueValue, nominalWeight);
-
-    if (isSelected)
-    {
-        // Add this event to the counter for selected signal events
-        m_signalSelectedDetVarRecoTrue.at(runId).at(detectorVariationParameter)->Fill(trueValue, recoValue, nominalWeight);
-    }
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::CrossSection::AddSelectedBackgroundEvent(const float recoValue, const float nominalWeight, const SystematicWeightsMap &systWeightsMap, const bool isOverlay)
-{
-    // Check the validity of the input map if systematic weights
-    if (!CrossSectionHelper::CheckSystematicWeightsMapDimensions(systWeightsMap, m_inputData.m_systUniverseSizesMap))
-        throw std::invalid_argument("CrossSection::AddSignalEvent - Input systematic weights map has the wrong dimensions!");
-
-    // Add this event in the nominal universe
-    m_backgroundSelectedNomReco->Fill(recoValue, nominalWeight);
-
-    // Add the event in each systematic universe
-    for (const auto &[name, nUniverses] : m_inputData.m_systUniverseSizesMap)
-    {
-        auto &backgroundSelectedRecoVect = m_backgroundSelectedReco.at(name);
-
-        const auto &weights = systWeightsMap.at(name);
-        for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
-        {
-            const auto universeWeight = weights.at(iUni);
-            backgroundSelectedRecoVect.at(iUni)->Fill(recoValue, nominalWeight * universeWeight);
-        }
-    }
-
-    // ATTN here we also count non-overlay events so they can be used when calculating the detector systematics
-    if (!isOverlay)
-    {
-        m_backgroundSelectedNonOverlayDetVarReco->Fill(recoValue, nominalWeight);
-    }
-
-    // Mark that the inputs have changed and any cached values need to be re-calculated
-    m_shouldResetCache = true;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::CrossSection::AddSelectedBackgroundOverlayEventDetVar(const float recoValue, const float nominalWeight, const std::string &runId, const std::string &detectorVariationParameter)
-{
-    // Check if this is a value run-ID / detector variation parameter
-    if (std::find(m_inputData.m_detVarParameters.begin(), m_inputData.m_detVarParameters.end(), std::pair<std::string, std::string> (runId, detectorVariationParameter)) == m_inputData.m_detVarParameters.end())
-    {
-        throw std::invalid_argument("CrossSection::AddSelectedBackgroundOverlayEventDetVar - Unknown detector variation parameter: " + detectorVariationParameter + " for run: " + runId);
-    }
-
-    // Add this event to the counter for selected background events
-    m_backgroundSelectedOverlayDetVarReco.at(runId).at(detectorVariationParameter)->Fill(recoValue, nominalWeight);
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::CrossSection::AddSelectedBNBDataEvent(const float recoValue)
-{
-    m_dataSelectedReco->Fill(recoValue);
-
-    // Mark that the inputs have changed and any cached values need to be re-calculated
-    m_shouldResetCache = true;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::GetEmptyFluxHist() const
-{
-    const std::string name = "xSecHist_" + std::to_string(m_histCount++);
-
-    // Get the binning of the flux histogram
-    const unsigned int nBins = m_inputData.m_flux->GetNbinsX();
-    std::vector<float> edges;
-
-    for (unsigned int iBin = 1; iBin <= nBins; ++iBin)
-    {
-        edges.push_back(m_inputData.m_flux->GetBinLowEdge(iBin));
-    }
-
-    // Add the last bin edge
-    edges.push_back(m_inputData.m_flux->GetBinLowEdge(nBins) + m_inputData.m_flux->GetBinWidth(nBins));
-
-    return std::make_shared<TH1F>(name.c_str(), "", edges.size() - 1, edges.data());
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::GetEmptyHist1D() const
-{
-    const std::string name = "xSecHist_" + std::to_string(m_histCount++);
-    return std::make_shared<TH1F>(name.c_str(), "", m_binEdges.size() - 1, m_binEdges.data());
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH2F> CrossSectionHelper::CrossSection::GetEmptyHist2D() const
-{
-    const std::string name = "xSecHist_" + std::to_string(m_histCount++);
-    return std::make_shared<TH2F>(name.c_str(), "", m_binEdges.size() - 1, m_binEdges.data(), m_binEdges.size() - 1, m_binEdges.data());
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-bool CrossSectionHelper::CrossSection::IsUnderOverflowBin(const unsigned int binIndex) const
-{
-    const auto nBins = m_binEdges.size() - 1;
-    if (binIndex > nBins)
-        throw std::out_of_range("CrossSection::IsUnderOverflowBin - input bin index " + std::to_string(binIndex) + " is out of range");
-
-    return ((binIndex == 1 && m_hasUnderflow) || (binIndex == nBins && m_hasOverflow));
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH2F> CrossSectionHelper::CrossSection::GetSmearingMatrix(const std::shared_ptr<TH2F> &signalSelectedRecoTrue) const
-{
-    // Check the input matrix has the correct dimensions
-    const auto nBins = m_binEdges.size() - 1;
-
-    if (static_cast<unsigned int>(signalSelectedRecoTrue->GetXaxis()->GetNbins()) != nBins ||
-        static_cast<unsigned int>(signalSelectedRecoTrue->GetYaxis()->GetNbins()) != nBins )
-    {
-        throw std::logic_error("CrossSection::GetSmearingMatrix - input matrix has the wrong dimensions");
-    }
-
-    auto smearingMatrix = this->GetEmptyHist2D();
-    for (unsigned int iTrue = 1; iTrue <= nBins; ++iTrue)
-    {
-        // Get the sum of the reco values for this true bin
-        float summedValue = 0.f;
-        for (unsigned int iReco = 1; iReco <= nBins; ++iReco)
-        {
-            summedValue += signalSelectedRecoTrue->GetBinContent(iTrue, iReco);
+            // Store the result
+            pReweightedFlux->SetBinContent(iBin, reweightedFlux);
         }
 
-        if (summedValue <= std::numeric_limits<float>::epsilon())
+        // Return the integrated flux
+        return this->GetIntegratedFlux(pReweightedFlux);
+
+    }, m_dimensions)
+{
+    // There should be one more bin edge than the number of bins
+    if (binEdges.size() != binValuesNominal.size() + 1)
+        throw std::invalid_argument("FluxReweightor::FluxReweightor - Number of bin edges doesn't match number of bins!");
+
+    // Fill the nominal flux histogram
+    for (unsigned int iBin = 1; iBin <= binValuesNominal.size(); ++iBin)
+    {
+        // ATTN here we shift the index by 1 as root enumerates bins from 1 and the vector enumerates from 0
+        m_pFluxNominal->SetBinContent(iBin, binValuesNominal.at(iBin - 1));
+    }
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+void CrossSectionHelper::FluxReweightor::AddEvent(const float trueNuEnergy, const float nominalWeight, const SystFloatMap &fluxWeights)
+{
+    // Make sure we have a valid input map
+    CrossSectionHelper::ValidateSystMap(fluxWeights, m_dimensions);
+
+    // Here we clear any cached values for the integrated flux variations because we are modifying the event rate spectra
+    m_getIntegratedFluxVariationFunction.ClearCache();
+
+    // Fill the nominal spectrum
+    m_pSpectrumNominal->Fill(trueNuEnergy, nominalWeight);
+
+    // Fill the universes
+    for (const auto &[paramName, weights] : fluxWeights)
+    {
+        auto &spectrumUniverses = m_spectrumVariations.at(paramName);
+
+        for (unsigned int iUni = 0; iUni < weights.size(); ++iUni)
         {
-            // ATTN if a true bin has no entries, then we have no information to work with
-            // Here we make the choice to give each reco bin an equal probability
-            const auto value = 1.f / static_cast<float>(nBins);
-            for (unsigned int iReco = 1; iReco <= nBins; ++iReco)
-            {
-                smearingMatrix->SetBinContent(iTrue, iReco, value);
-            }
-        }
-        else
-        {
-            // Normalise the by a constant factor bins such that the columns add to unity
-            const auto norm = 1.f / summedValue;
-
-            for (unsigned int iReco = 1; iReco <= nBins; ++iReco)
-            {
-                const auto value = signalSelectedRecoTrue->GetBinContent(iTrue, iReco) * norm;
-                smearingMatrix->SetBinContent(iTrue, iReco, value);
-            }
-        }
-    }
-
-    return smearingMatrix;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::Smear(const std::shared_ptr<TH1F> &values, const std::shared_ptr<TH2F> &smearingMatrix) const
-{
-    // Check the input objects have the correct dimensions
-    const auto nBins = m_binEdges.size() - 1;
-
-    if (static_cast<unsigned int>(values->GetXaxis()->GetNbins()) != nBins)
-        throw std::logic_error("CrossSection::Smear - input values has the wrong dimensions");
-
-    if (static_cast<unsigned int>(smearingMatrix->GetXaxis()->GetNbins()) != nBins ||
-        static_cast<unsigned int>(smearingMatrix->GetYaxis()->GetNbins()) != nBins)
-    {
-        throw std::logic_error("CrossSection::Smear - input matrix has the wrong dimensions");
-    }
-
-
-    auto smearedValues = this->GetEmptyHist1D();
-
-    for (unsigned int iReco = 1; iReco <= nBins; ++iReco)
-    {
-        // Smear the true values in this reco bin
-        float smearedValue = 0.f;
-        for (unsigned int iTrue = 1; iTrue <= nBins; ++iTrue)
-        {
-            smearedValue += values->GetBinContent(iTrue) * smearingMatrix->GetBinContent(iTrue, iReco);
-        }
-
-        smearedValues->SetBinContent(iReco, smearedValue);
-    }
-
-    return smearedValues;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::GetSmearedEfficiency(const std::shared_ptr<TH2F> &signalSelectedRecoTrue, const std::shared_ptr<TH1F> &signalAllTrue) const
-{
-    const auto smearingMatrix = this->GetSmearingMatrix(signalSelectedRecoTrue);
-    return this->GetSmearedEfficiency(smearingMatrix, signalSelectedRecoTrue, signalAllTrue);
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::GetSmearedEfficiency(const std::shared_ptr<TH2F> &smearingMatrix, const std::shared_ptr<TH2F> &signalSelectedRecoTrue, const std::shared_ptr<TH1F> &signalAllTrue) const
-{
-    // Check the input objects have the correct dimensions
-    const auto nBins = m_binEdges.size() - 1;
-
-    if (static_cast<unsigned int>(signalAllTrue->GetXaxis()->GetNbins()) != nBins)
-        throw std::logic_error("CrossSection::GetSmearedEfficiency - input object signalAllTrue has the wrong dimensions");
-
-    if (static_cast<unsigned int>(smearingMatrix->GetXaxis()->GetNbins()) != nBins ||
-        static_cast<unsigned int>(smearingMatrix->GetYaxis()->GetNbins()) != nBins)
-    {
-        throw std::logic_error("CrossSection::GetSmearedEfficiency - input smearing matrix has the wrong dimensions");
-    }
-
-    if (static_cast<unsigned int>(signalSelectedRecoTrue->GetXaxis()->GetNbins()) != nBins ||
-        static_cast<unsigned int>(signalSelectedRecoTrue->GetYaxis()->GetNbins()) != nBins)
-    {
-        throw std::logic_error("CrossSection::GetSmearedEfficiency - input signalSelectedRecoTrue matrix has the wrong dimensions");
-    }
-
-    // Get the total selected events per true bin
-    auto signalSelectedTrue = this->GetEmptyHist1D();
-    for (unsigned int iTrue = 1; iTrue <= nBins; ++iTrue)
-    {
-        // Get the sum of the reco values for this true bin
-        float summedValue = 0.f;
-        for (unsigned int iReco = 1; iReco <= nBins; ++iReco)
-        {
-            summedValue += signalSelectedRecoTrue->GetBinContent(iTrue, iReco);
-        }
-
-        signalSelectedTrue->SetBinContent(iTrue, summedValue);
-    }
-
-    // Now smear the numerator and denominators of the efficiency
-    const auto signalSelectedSmearedReco = this->Smear(signalSelectedTrue, smearingMatrix);
-    const auto signalAllSmearedReco = this->Smear(signalAllTrue, smearingMatrix);
-
-    // Get the smeared efficiency
-    auto effSmearedReco = this->GetEmptyHist1D();
-    for (unsigned int iReco = 1; iReco <= nBins; ++iReco)
-    {
-        // Skip under/overflow bins
-        if (this->IsUnderOverflowBin(iReco))
-            continue;
-
-        const auto numerator = signalSelectedSmearedReco->GetBinContent(iReco);
-        const auto denominator = signalAllSmearedReco->GetBinContent(iReco);
-
-        const auto isDenominatorZero = denominator <= std::numeric_limits<float>::epsilon();
-
-        // ATTN here we check a zero denominator - in this we set a dummy efficiency of -inf
-        const auto value = isDenominatorZero ? -std::numeric_limits<float>::max() : (numerator / denominator);
-        effSmearedReco->SetBinContent(iReco, value);
-    }
-
-    return effSmearedReco;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::CrossSection::ClearCache()
-{
-    m_crossSectionCache.clear();
-    m_smearingMatrixCache.clear();
-    m_shouldResetCache = false;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::GetCachedCrossSection(const std::string &systParameter, const unsigned int universeIndex)
-{
-    // Check that we have an entry for the requested systematic paramter
-    const auto iter = m_inputData.m_systUniverseSizesMap.find(systParameter);
-    if (iter == m_inputData.m_systUniverseSizesMap.end())
-        throw std::invalid_argument("CrossSection::GetCachedCrossSection - unknown parameter: \"" + systParameter + "\"");
-
-    // Get the number of universes
-    const auto nUniverses = iter->second;
-    if (universeIndex >= nUniverses)
-        throw std::range_error("CrossSection::GetCachedCrossSection - Universe index for parameter: \"" + systParameter + "\" is out of range");
-
-    // Reset the cached cross-section if required
-    if (m_shouldResetCache)
-        this->ClearCache();
-
-    // Get the cached cross-section (if it exists)
-    const auto paramIter = m_crossSectionCache.find(systParameter);
-    if (paramIter != m_crossSectionCache.end())
-    {
-        const auto universeIter = paramIter->second.find(universeIndex);
-        if (universeIter != paramIter->second.end())
-            return universeIter->second;
-    }
-
-    // We don't have an entry for this cross-section, so make it!
-
-    // Determine if this is a flux parameter and if so, reweight the total flux
-    const auto isFluxParam = (std::find(m_inputData.m_fluxParameters.begin(), m_inputData.m_fluxParameters.end(), systParameter) != m_inputData.m_fluxParameters.end());
-
-    const auto totalFlux = CrossSectionHelper::GetTotalFlux(isFluxParam
-            ? this->GetReweightedFlux(m_allEventsNuEnergyTrue.at(systParameter).at(universeIndex))
-            : m_inputData.m_flux);
-
-    // Get the smeared efficiency distribution and cache the smearing matrix if required
-    const auto smearingMatrix = this->GetCachedSmearingMatrix(systParameter, universeIndex);
-    const auto smearedEff = this->GetSmearedEfficiency(smearingMatrix, m_signalSelectedRecoTrue.at(systParameter).at(universeIndex), m_signalAllTrue.at(systParameter).at(universeIndex));
-
-    // Get the cross-section and cache it
-    auto &cachedXSec = m_crossSectionCache[systParameter][universeIndex];
-    cachedXSec = this->GetCrossSection(                                         // Get the total cross-section using:
-        m_dataSelectedReco,                                                     //   - Selected BNB data
-        m_backgroundSelectedReco.at(systParameter).at(universeIndex),           //   - Backgrounds in this universe
-        smearedEff, totalFlux);                                                 //   - Smeared efficiency and flux in this universe
-
-    return cachedXSec;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH2F> CrossSectionHelper::CrossSection::GetCachedSmearingMatrix(const std::string &systParameter, const unsigned int universeIndex)
-{
-    // Check that we have an entry for the requested systematic paramter
-    const auto iter = m_inputData.m_systUniverseSizesMap.find(systParameter);
-    if (iter == m_inputData.m_systUniverseSizesMap.end())
-        throw std::invalid_argument("CrossSection::GetCachedSmearingMatrix - unknown parameter: \"" + systParameter + "\"");
-
-    // Get the number of universes
-    const auto nUniverses = iter->second;
-    if (universeIndex >= nUniverses)
-        throw std::range_error("CrossSection::GetCachedSmearingMatrix - Universe index for parameter: \"" + systParameter + "\" is out of range");
-
-    // Reset the cached cross-section if required
-    if (m_shouldResetCache)
-        this->ClearCache();
-
-    // Get the cached smearing matrix (if it exists)
-    const auto paramIter = m_smearingMatrixCache.find(systParameter);
-    if (paramIter != m_smearingMatrixCache.end())
-    {
-        const auto universeIter = paramIter->second.find(universeIndex);
-        if (universeIter != paramIter->second.end())
-            return universeIter->second;
-    }
-
-    // We don't have an entry for this smearing matrix, so make it!
-    auto &cachedSmearingMatrix = m_smearingMatrixCache[systParameter][universeIndex];
-    cachedSmearingMatrix = this->GetSmearingMatrix(m_signalSelectedRecoTrue.at(systParameter).at(universeIndex));
-
-    return cachedSmearingMatrix;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::GetCrossSection(const std::shared_ptr<TH1F> &selected, const std::shared_ptr<TH1F> &background, const std::shared_ptr<TH1F> &smearedEff, const float totalFlux) const
-{
-    const auto normFactor = 1.f / (totalFlux * m_inputData.m_exposurePOT * m_inputData.m_nTargets);
-
-    // Check the input objects have the correct dimensions
-    const auto nBins = m_binEdges.size() - 1;
-
-    if (static_cast<unsigned int>(selected->GetXaxis()->GetNbins()) != nBins)
-        throw std::logic_error("CrossSection::GetCrossSection - input selected distribution has the wrong dimensions");
-
-    if (static_cast<unsigned int>(background->GetXaxis()->GetNbins()) != nBins)
-        throw std::logic_error("CrossSection::GetCrossSection - input background distribution has the wrong dimensions");
-
-    if (static_cast<unsigned int>(smearedEff->GetXaxis()->GetNbins()) != nBins)
-        throw std::logic_error("CrossSection::GetCrossSection - input smeared efficiency distribution has the wrong dimensions");
-
-    auto crossSection = this->GetEmptyHist1D();
-    for (unsigned int iReco = 1; iReco <= nBins; ++iReco)
-    {
-        // Skip under/overflow bins
-        if (this->IsUnderOverflowBin(iReco))
-            continue;
-
-        // In some universes, the efficiency may be incalculable (and so a negative dummy value is used) - here we use a dummy value again
-        const auto efficiency = smearedEff->GetBinContent(iReco);
-        if (std::abs(std::abs(efficiency) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon() ||
-            efficiency <= std::numeric_limits<float>::epsilon())
-        {
-            crossSection->SetBinContent(iReco, -std::numeric_limits<float>::max());
-            continue;
-        }
-
-        const auto binWidth = m_scaleByBinWidth ? (m_binEdges.at(iReco) - m_binEdges.at(iReco - 1)) : 1.f;
-        const auto value = normFactor * (selected->GetBinContent(iReco) - background->GetBinContent(iReco)) / (efficiency * binWidth);
-
-        crossSection->SetBinContent(iReco, value);
-    }
-
-    return crossSection;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH2F> CrossSectionHelper::CrossSection::GetSmearingMatrix() const
-{
-    return this->GetSmearingMatrix(m_signalSelectedNomRecoTrue);
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::GetCrossSection() const
-{
-    const auto smearedEff = this->GetSmearedEfficiency(m_signalSelectedNomRecoTrue, m_signalAllNomTrue);
-    const auto totalFlux = CrossSectionHelper::GetTotalFlux(m_inputData.m_flux);
-    return this->GetCrossSection(m_dataSelectedReco, m_backgroundSelectedNomReco, smearedEff, totalFlux);
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::GetCrossSectionStatUncertainty() const
-{
-    // Setup the output histogram
-    const auto nBins = m_binEdges.size() - 1;
-    const auto nAnalysisBins = nBins - (m_hasUnderflow ? 1 : 0) - (m_hasOverflow ? 1 : 0);
-    const std::string name = "xSecStatErr_" + std::to_string(m_histCount++);
-    auto statErrHist = std::make_shared<TH1F>(name.c_str(), "", nAnalysisBins, 0, nAnalysisBins);
-
-    const auto smearedEff = this->GetSmearedEfficiency(m_signalSelectedNomRecoTrue, m_signalAllNomTrue);
-    const auto totalFlux = CrossSectionHelper::GetTotalFlux(m_inputData.m_flux);
-    const auto normFactor = 1.f / (totalFlux * m_inputData.m_exposurePOT * m_inputData.m_nTargets);
-
-    for (unsigned int iBin = 1u; iBin <= nBins; ++iBin)
-    {
-        if (this->IsUnderOverflowBin(iBin))
-            continue;
-
-        const auto binWidth = m_scaleByBinWidth ? (m_binEdges.at(iBin) - m_binEdges.at(iBin - 1)) : 1.f;
-        const auto nDataEvents = m_dataSelectedReco->GetBinContent(iBin);
-        const auto efficiency = smearedEff->GetBinContent(iBin);
-
-        if (std::abs(std::abs(efficiency) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon() ||
-            efficiency <= std::numeric_limits<float>::epsilon())
-        {
-            throw std::logic_error("CrossSection::GetCrossSectionStatUncertainty - Found bin in which smeared efficiency couldn't be calculated!");
-        }
-
-        // Get the uncertainty on the number of data events
-        const auto nDataEventsErr = AnalysisHelper::GetCountUncertainty(nDataEvents);
-
-        // Get the corresponding uncertainty on the cross-section
-        // Cross-section = normFactor * (nDataEvents - nBackgroundsEvents) / (efficiency * binWidth)
-        const auto uncertainty = nDataEventsErr * normFactor / (efficiency * binWidth);
-
-        statErrHist->SetBinContent(iBin - (m_hasUnderflow ? 1 : 0), uncertainty);
-    }
-
-    return statErrHist;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::map< std::string, std::vector< std::tuple<unsigned int, unsigned int, std::shared_ptr<TGraph> > > > CrossSectionHelper::CrossSection::GetCrossSectionBinScatterPlots()
-{
-    std::map< std::string, std::vector< std::tuple<unsigned int, unsigned int, std::shared_ptr<TGraph> > > > outputMap;
-
-    for (const auto &[param, nUniverses] : m_inputData.m_systUniverseSizesMap)
-    {
-        outputMap.emplace(param, this->GetCrossSectionBinScatterPlots(param));
-    }
-
-    return outputMap;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::vector< std::tuple<unsigned int, unsigned int, std::shared_ptr<TGraph> > > CrossSectionHelper::CrossSection::GetCrossSectionBinScatterPlots(const std::string &systParameter)
-{
-    // Check that we have an entry for the requested systematic paramter
-    const auto iter = m_inputData.m_systUniverseSizesMap.find(systParameter);
-    if (iter == m_inputData.m_systUniverseSizesMap.end())
-        throw std::invalid_argument("CrossSection::GetCrossSectionBinScatterPlots - unknown parameter: \"" + systParameter + "\"");
-
-    const auto nUniverses = iter->second;
-
-    // Setup the output vector
-    std::vector< std::tuple<unsigned int, unsigned int, std::shared_ptr<TGraph> > > outputVect;
-
-    // Loop over all pairs of bins
-    const auto nBins = m_binEdges.size() - 1;
-    for (unsigned int iBin = 1u; iBin <= nBins; ++iBin)
-    {
-        if (this->IsUnderOverflowBin(iBin))
-            continue;
-
-        for (unsigned int jBin = 1u; jBin <= iBin; ++jBin)
-        {
-            if (this->IsUnderOverflowBin(jBin))
-                continue;
-
-            // Get the cross-section in each universe from the cache - or add it if required
-            std::vector<float> iValues, jValues;
-            float minValue = std::numeric_limits<float>::max();
-            float maxValue = -std::numeric_limits<float>::max();
-
-            for (unsigned int iUni = 0u; iUni < nUniverses; ++iUni)
-            {
-                // Get the values of the cross-section in the current pair of bins and make sure they are valid
-                const auto crossSection = this->GetCachedCrossSection(systParameter, iUni);
-
-                const float iValue = crossSection->GetBinContent(iBin);
-                if (std::abs(std::abs(iValue) - std::numeric_limits<float>::max()) < std::numeric_limits<float>::epsilon())
-                    continue;
-
-                const float jValue = crossSection->GetBinContent(jBin);
-                if (std::abs(std::abs(jValue) - std::numeric_limits<float>::max()) < std::numeric_limits<float>::epsilon())
-                    continue;
-
-                // Store these values
-                iValues.push_back(iValue);
-                jValues.push_back(jValue);
-
-                // Get the minimum and maximum values
-                minValue = std::min(minValue, std::min(iValue, jValue));
-                maxValue = std::max(maxValue, std::max(iValue, jValue));
-            }
-
-            // Store the values in a scatter plot
-            const auto nValues = iValues.size();
-            auto pGraph = std::make_shared<TGraph>(nValues, iValues.data(), jValues.data());
-
-            // Set the range of the graph to ensure it's square
-            pGraph->GetXaxis()->SetLimits(minValue, maxValue);
-            pGraph->GetYaxis()->SetLimits(minValue, maxValue);
-
-            outputVect.emplace_back(iBin, jBin, pGraph);
+            const auto universeWeight = nominalWeight * weights.at(iUni);
+            spectrumUniverses.at(iUni)->Fill(trueNuEnergy, universeWeight);
         }
     }
-
-    return outputVect;
 }
 
 // -----------------------------------------------------------------------------------------------------------------------------------------
 
-CrossSectionHelper::CovarianceBiasPair CrossSectionHelper::CrossSection::GetCrossSectionCovarianceMatrix(const std::string &systParameter)
+std::shared_ptr<TH1F> CrossSectionHelper::FluxReweightor::GetNominalFlux() const
 {
-    // Check that we have an entry for the requested systematic paramter
-    const auto iter = m_inputData.m_systUniverseSizesMap.find(systParameter);
-    if (iter == m_inputData.m_systUniverseSizesMap.end())
-        throw std::invalid_argument("CrossSection::GetCrossSectionCovarianceMatrix - unknown parameter: \"" + systParameter + "\"");
-
-    // Get the number of universes
-    const auto nUniverses = iter->second;
-    if (nUniverses == 0)
-        throw std::logic_error("CrossSection::GetCrossSectionCovarianceMatrix - No universes for parameter: \"" + systParameter + "\"");
-
-    // Get the nominal cross section
-    const auto crossSectionNom = this->GetCrossSection();
-
-    // Get the cross-section in each universe and the mean cross-section over all universes
-    std::vector< std::shared_ptr<TH1F> > crossSections;
-
-    auto nUniversesPerBin = this->GetEmptyHist1D();
-    auto summedCrossSection = this->GetEmptyHist1D();
-    const auto nBins = m_binEdges.size() - 1;
-
-    //// BEGIN DEBUG
-    std::cout << "Getting covariance matrix for parameter: " << systParameter << std::endl;
-    //// END DEBUG
-
-    for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
-    {
-        //// BEGIN DEBUG
-        AnalysisHelper::PrintLoadingBar(iUni, nUniverses);
-        //// END DEBUG
-
-        // Get the cross-section from the cache (or add it if we haven't calculated this cross section yet)
-        const auto crossSection = this->GetCachedCrossSection(systParameter, iUni);
-        crossSections.push_back(crossSection);
-
-        // Add this to the sum for each bin
-        for (unsigned int iBin = 1; iBin <= nBins; ++iBin)
-        {
-            if (this->IsUnderOverflowBin(iBin))
-                continue;
-
-            // Skip dummy bins in which the cross-section couldn't be extracted
-            const auto binValue = crossSection->GetBinContent(iBin);
-            if (std::abs(std::abs(binValue) - std::numeric_limits<float>::max()) < std::numeric_limits<float>::epsilon())
-                continue;
-
-            summedCrossSection->SetBinContent(iBin, summedCrossSection->GetBinContent(iBin) + binValue);
-            nUniversesPerBin->SetBinContent(iBin, nUniversesPerBin->GetBinContent(iBin) + 1.f);
-        }
-    }
-
-    // Get the mean cross-section
-    auto meanCrossSection = this->GetEmptyHist1D();
-    for (unsigned int iBin = 1; iBin <= nBins; ++iBin)
-    {
-        if (this->IsUnderOverflowBin(iBin))
-            continue;
-
-        const auto numerator = summedCrossSection->GetBinContent(iBin);
-        const auto denominator = nUniversesPerBin->GetBinContent(iBin);
-
-        // If the denominator is zero then use a dummy value
-        const auto isDemoninatorZero = (denominator <= std::numeric_limits<float>::epsilon());
-        const auto mean = isDemoninatorZero ? -std::numeric_limits<float>::max() : (numerator / denominator);
-        meanCrossSection->SetBinContent(iBin, mean);
-    }
-
-    // Get the covairance matrix and bias vector
-    const auto nAnalysisBins = nBins - (m_hasUnderflow ? 1 : 0) - (m_hasOverflow ? 1 : 0);
-    const std::string covName = "xSecCov_" + std::to_string(m_histCount++);
-    const std::string biasName = "xSecBias_" + std::to_string(m_histCount++);
-    CovarianceBiasPair pair (
-        std::make_shared<TH2F>(covName.c_str(), "", nAnalysisBins, 0, nAnalysisBins, nAnalysisBins, 0, nAnalysisBins),
-        std::make_shared<TH1F>(biasName.c_str(), "", nAnalysisBins, 0, nAnalysisBins)
-    );
-
-    auto &covarianceMatrix = pair.first;
-    auto &biasVector = pair.second;
-
-    for (unsigned int iBin = 1; iBin <= nBins; ++iBin)
-    {
-        if (this->IsUnderOverflowBin(iBin))
-            continue;
-
-        const auto meanI = meanCrossSection->GetBinContent(iBin);
-        const auto nomI = crossSectionNom->GetBinContent(iBin);
-
-        // Check for dummy bins and warn if they exists
-        if (std::abs(std::abs(nomI) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon())
-        {
-            std::cout << "WARNING - Bin: " << iBin << " of nominal cross-section is invalid" << std::endl;
-            continue;
-        }
-
-        if (std::abs(std::abs(meanI) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon())
-        {
-            std::cout << "WARNING - Bin: " << iBin << " of has an incalculable mean value over all universe variations" << std::endl;
-            continue;
-        }
-
-        // Set the bias vector element
-        const auto bias = meanI - nomI;
-        const auto iBinOutput = iBin - (m_hasUnderflow ? 1 : 0);
-        biasVector->SetBinContent(iBinOutput, bias);
-
-        for (unsigned int jBin = 1; jBin <= nBins; ++jBin)
-        {
-            if (this->IsUnderOverflowBin(jBin))
-                continue;
-
-            const auto meanJ = meanCrossSection->GetBinContent(jBin);
-            const auto nomJ = crossSectionNom->GetBinContent(jBin);
-
-            // Skip dummy bins
-            if (std::abs(std::abs(nomJ) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon())
-                continue;
-
-            if (std::abs(std::abs(meanJ) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon())
-                continue;
-
-            float numerator = 0.f;
-            float denominator = 0.f;
-            for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
-            {
-                const auto &crossSection = crossSections.at(iUni);
-                const auto sigmaI = crossSection->GetBinContent(iBin);
-                const auto sigmaJ = crossSection->GetBinContent(jBin);
-
-                // Skip dummy bins
-                if (std::abs(std::abs(sigmaI) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon())
-                    continue;
-
-                if (std::abs(std::abs(sigmaJ) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon())
-                    continue;
-
-                numerator   += (sigmaI - meanI) * (sigmaJ - meanJ);
-                denominator += 1.f;
-            }
-
-            if (denominator <= std::numeric_limits<float>::epsilon())
-            {
-                throw std::logic_error("CrossSection::GetCrossSectionCovarianceMatrix - For parameter: \"" + systParameter + "\", " +
-                    "can't calculate covariance in bin: " + std::to_string(iBin) + ", " + std::to_string(jBin));
-            }
-
-            // Set the covariance matrix element
-            const auto covariance = numerator / denominator;
-            const auto jBinOutput = jBin - (m_hasUnderflow ? 1 : 0);
-            covarianceMatrix->SetBinContent(iBinOutput, jBinOutput, covariance);
-        }
-    }
-
-    return pair;
+    return m_pFluxNominal;
 }
 
 // -----------------------------------------------------------------------------------------------------------------------------------------
 
-CrossSectionHelper::CovarianceBiasPair CrossSectionHelper::CrossSection::GetCrossSectionCovarianceMatrixDetVar(const std::string &runId, const std::string &detVarParam)
-{
-    // Check if this is a valid run-ID / detector variation parameter
-    if (std::find(m_inputData.m_detVarParameters.begin(), m_inputData.m_detVarParameters.end(), std::pair<std::string, std::string> (runId, detVarParam)) == m_inputData.m_detVarParameters.end())
-    {
-        throw std::invalid_argument("CrossSection::GetCrossSectionCovarianceMatrixDetVar - Unknown detector variation parameter: " + detVarParam + " for run: " + runId);
-    }
-
-    // Get the total flux
-    const auto totalFlux = CrossSectionHelper::GetTotalFlux(m_inputData.m_flux);
-
-    // Get the smeared efficiency in the central value sample and the sample for this detector parameter
-    const auto smearedEffCV = this->GetSmearedEfficiency(m_signalSelectedDetVarRecoTrue.at(runId).at(m_cvString), m_signalAllDetVarTrue.at(runId).at(m_cvString));
-    const auto smearedEff = this->GetSmearedEfficiency(m_signalSelectedDetVarRecoTrue.at(runId).at(detVarParam), m_signalAllDetVarTrue.at(runId).at(detVarParam));
-
-    // Get the backgrounds in the central value sample and the sample for this detector parameter. Here we add the backgrounds from the
-    // overlay detector variation samples, to those from other sources - e.g. dirt or EXT.
-    auto backgroundsCV = this->GetEmptyHist1D();
-    auto backgrounds = this->GetEmptyHist1D();
-
-    // Overlay backgrounds (from detector variation samples)
-    backgroundsCV->Add(m_backgroundSelectedOverlayDetVarReco.at(runId).at(m_cvString).get());
-    backgrounds->Add(m_backgroundSelectedOverlayDetVarReco.at(runId).at(detVarParam).get());
-
-    // Non overlay backgrounds (dirt / EXT)
-    backgroundsCV->Add(m_backgroundSelectedNonOverlayDetVarReco.get());
-    backgrounds->Add(m_backgroundSelectedNonOverlayDetVarReco.get());
-
-    //// BEGIN DEBUG
-    // Print the difference between the backgrounsd in the central value and nominal samples
-    std::cout << "BACKGROUNDS" << std::endl;
-    for (unsigned int iBin = 1u; iBin <= static_cast<unsigned int>(backgroundsCV->GetNbinsX()); ++iBin)
-    {
-        std::cout << "Bin: " << iBin << std::endl;
-        std::cout << "  - CV:  " << backgroundsCV->GetBinContent(iBin) << std::endl;
-        std::cout << "  - Var: " << backgrounds->GetBinContent(iBin) << std::endl;
-        std::cout << "  - Nom: " << m_backgroundSelectedNomReco->GetBinContent(iBin) << std::endl;
-    }
-    //// END DEBUG
-
-    // Get the cross-section in the central value sample and the sample for this detector parameter
-    const auto crossSectionCV = this->GetCrossSection(m_dataSelectedReco, backgroundsCV, smearedEffCV, totalFlux);
-    const auto crossSection = this->GetCrossSection(m_dataSelectedReco, backgrounds, smearedEff, totalFlux);
-
-    // Now get the nominal cross-section (using the normal overlays sample)
-    const auto crossSectionNom = this->GetCrossSection();
-
-    // Build the covariance matrix and bias vector
-    const auto nBins = m_binEdges.size() - 1;
-    const auto nAnalysisBins = nBins - (m_hasUnderflow ? 1 : 0) - (m_hasOverflow ? 1 : 0);
-
-    const std::string covName = "xSecCov_" + std::to_string(m_histCount++);
-    const std::string biasName = "xSecBias_" + std::to_string(m_histCount++);
-    CovarianceBiasPair pair (
-        std::make_shared<TH2F>(covName.c_str(), "", nAnalysisBins, 0, nAnalysisBins, nAnalysisBins, 0, nAnalysisBins),
-        std::make_shared<TH1F>(biasName.c_str(), "", nAnalysisBins, 0, nAnalysisBins)
-    );
-
-    // ATTN here the covariance matrix shall remain zero by construction, we will instead treat the detector variation entirely as a bias
-    auto &biasVector = pair.second;
-
-    for (unsigned int iBin = 1; iBin <= nBins; ++iBin)
-    {
-        if (this->IsUnderOverflowBin(iBin))
-            continue;
-
-        const auto xsec = crossSection->GetBinContent(iBin);
-        const auto xsecCV = crossSectionCV->GetBinContent(iBin);
-        const auto xsecNom = crossSectionNom->GetBinContent(iBin);
-
-        // Check for dummy bins and warn if they exists
-        if (std::abs(std::abs(xsec) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon())
-        {
-            std::cout << "WARNING - Bin: " << iBin << " of cross-section for detector parameter: " + detVarParam + " is invalid"  << std::endl;
-            continue;
-        }
-
-        if (std::abs(std::abs(xsecCV) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon())
-        {
-            std::cout << "WARNING - Bin: " << iBin << " of cross-section for central value of detector variations in run: " + runId + " is invalid"  << std::endl;
-            continue;
-        }
-
-        if (std::abs(std::abs(xsecNom) - std::numeric_limits<float>::max()) <= std::numeric_limits<float>::epsilon())
-        {
-            std::cout << "WARNING - Bin: " << iBin << " of cross-section in nominal sample is invalid"  << std::endl;
-            continue;
-        }
-
-        // Get the scaling factor to translate the central value of the detector variations to the nominal simulation
-        if (std::abs(xsecCV) <= std::numeric_limits<float>::epsilon())
-            throw std::logic_error("CrossSection::GetCrossSectionCovarianceMatrixDetVar - Central value cross-section in bin: " + std::to_string(iBin) + " is zero");
-
-        const auto cvToNomFactor = xsecNom / xsecCV;
-
-        // Get the bias vector element
-        // Here we find the difference between the cross-section as calculated with and without the detector variation applied. The central
-        // value sample contains the same events as those in the detector variation sample. We then scale this difference by the ratio of
-        // the cross-section in the nominal to the central value samples.
-        const auto bias = (xsec - xsecCV) * cvToNomFactor;
-
-        //// BEGIN DEBUG
-        std::cout << "Detector parameter: " << detVarParam << std::endl;
-        std::cout << "  - Bin: " << iBin << std::endl;
-        std::cout << "  - Variation: " << xsec << std::endl;
-        std::cout << "  - CV (" << runId << "): " << xsecCV << std::endl;
-        std::cout << "  - Nominal: " << xsecNom << std::endl;
-        std::cout << "  - CV-to-Nom factor: " << cvToNomFactor << std::endl;
-        std::cout << "  - Bias: " << bias << std::endl;
-        //// END DEBUG
-
-        // Store the result
-        const auto iBinOutput = iBin - (m_hasUnderflow ? 1 : 0);
-        biasVector->SetBinContent(iBinOutput, bias);
-    }
-
-    return pair;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-CrossSectionHelper::CovarianceBiasPair CrossSectionHelper::CrossSection::GetSmearingMatrixCovarianceMatrix(const std::string &systParameter)
-{
-    // Check that we have an entry for the requested systematic paramter
-    const auto iter = m_inputData.m_systUniverseSizesMap.find(systParameter);
-    if (iter == m_inputData.m_systUniverseSizesMap.end())
-        throw std::invalid_argument("CrossSection::GetSmearingMatrixCovarianceMatrix - unknown parameter: \"" + systParameter + "\"");
-
-    // Get the number of universes
-    const auto nUniverses = iter->second;
-    if (nUniverses == 0)
-        throw std::logic_error("CrossSection::GetSmearingMatrixCovarianceMatrix - No universes for parameter: \"" + systParameter + "\"");
-
-    // Setup the covairance matrix and bias vector (here we flatten the bins of the smearing matrix into 1D)
-    const auto nBins = m_binEdges.size() - 1;
-    const auto nSmearingBins = nBins*nBins;
-
-    const std::string covName = "xSecSmearingCov_" + std::to_string(m_histCount++);
-    const std::string biasName = "xSecSmearingBias_" + std::to_string(m_histCount++);
-    CovarianceBiasPair pair (
-        std::make_shared<TH2F>(covName.c_str(), "", nSmearingBins, 0, nSmearingBins, nSmearingBins, 0, nSmearingBins),
-        std::make_shared<TH1F>(biasName.c_str(), "", nSmearingBins, 0, nSmearingBins)
-    );
-
-    auto &covarianceMatrix = pair.first;
-    auto &biasVector = pair.second;
-
-    //// BEGIN DEBUG
-    std::cout << "Getting mean smearing matrix for parameter: " << systParameter << std::endl;
-    //// END DEBUG
-
-    // Get the mean smearing matrix
-    auto meanSmearingMatrix = this->GetEmptyHist2D();
-    for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
-    {
-        //// BEGIN DEBUG
-        AnalysisHelper::PrintLoadingBar(iUni, nUniverses);
-        //// END DEBUG
-
-        // Get the smearing matrix from the cache (or add it if we haven't calculated this smearing matrix yet)
-        const auto smearingMatrix = this->GetCachedSmearingMatrix(systParameter, iUni);
-        meanSmearingMatrix->Add(smearingMatrix.get());
-    }
-    const auto universeNormalisation = 1.f / static_cast<float>(nUniverses);
-    meanSmearingMatrix->Scale(universeNormalisation);
-
-    // Get the nominal smearing matrix
-    const auto nomSmearingMatrix = this->GetSmearingMatrix();
-
-    // Flatten the mean and nominal smearing matrices
-    std::vector<float> meanSmearingMatrixFlat, nomSmearingMatrixFlat;
-    for (unsigned int iTrue = 1u; iTrue <= nBins; ++iTrue)
-    {
-        for (unsigned int iReco = 1u; iReco <= nBins; ++iReco)
-        {
-            meanSmearingMatrixFlat.push_back(meanSmearingMatrix->GetBinContent(iTrue, iReco));
-            nomSmearingMatrixFlat.push_back(nomSmearingMatrix->GetBinContent(iTrue, iReco));
-        }
-    }
-
-    // Get the bias vector
-    for (unsigned int iFlat = 0u; iFlat < nSmearingBins; ++iFlat)
-    {
-        const auto bias = meanSmearingMatrixFlat.at(iFlat) - nomSmearingMatrixFlat.at(iFlat);
-        biasVector->SetBinContent(iFlat + 1, bias); // +1 to account for root enumerating bins from 1
-    }
-
-    //// BEGIN DEBUG
-    std::cout << "Getting covariance matrix of smearing matrix for parameter: " << systParameter << std::endl;
-    //// END DEBUG
-
-    // Get the covariance matrix
-    for (unsigned int iUni = 0u; iUni < nUniverses; ++iUni)
-    {
-        //// BEGIN DEBUG
-        AnalysisHelper::PrintLoadingBar(iUni, nUniverses);
-        //// END DEBUG
-
-        // Get the smearing matrix from the cache
-        const auto smearingMatrix = this->GetCachedSmearingMatrix(systParameter, iUni);
-
-        // Flatten the smearing matrix
-        std::vector<float> smearingMatrixFlat;
-        for (unsigned int iTrue = 1u; iTrue <= nBins; ++iTrue)
-        {
-            for (unsigned int iReco = 1u; iReco <= nBins; ++iReco)
-            {
-                smearingMatrixFlat.push_back(smearingMatrix->GetBinContent(iTrue, iReco));
-            }
-        }
-
-        // Loop over the flattened indices and add to the covariance matrix
-        for (unsigned int iFlat = 0u; iFlat < nSmearingBins; ++iFlat)
-        {
-            for (unsigned int jFlat = 0u; jFlat <= iFlat; ++jFlat)
-            {
-                const auto currentValue = covarianceMatrix->GetBinContent(iFlat + 1, jFlat + 1); // +1 to account for root enumerating bins from 1
-                const auto newValue = currentValue + (
-                        (smearingMatrixFlat.at(iFlat) - meanSmearingMatrixFlat.at(iFlat)) *
-                        (smearingMatrixFlat.at(jFlat) - meanSmearingMatrixFlat.at(jFlat)));
-
-                // The matrix is symmetric so we can switch iFlat -> jFlat
-                covarianceMatrix->SetBinContent(iFlat + 1, jFlat + 1, newValue);
-                covarianceMatrix->SetBinContent(jFlat + 1, iFlat + 1, newValue);
-            }
-        }
-    }
-    covarianceMatrix->Scale(universeNormalisation);
-
-    return pair;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::CrossSection::GetReweightedFlux(const std::shared_ptr<TH1F> &nuEnergyUniverse) const
-{
-    // Check the input histogram has the right number of bins
-    const unsigned int nBins = m_inputData.m_flux->GetNbinsX();
-    if (static_cast<unsigned int>(nuEnergyUniverse->GetNbinsX()) != nBins)
-        throw std::invalid_argument("CrossSection::GetReweightedFlux - Input distribution has the wrong number of bins");
-
-    // Get an empty histogram to fill with the re-weighted distribution
-    auto reweightedFlux = this->GetEmptyFluxHist();
-
-    // Reweight the flux
-    for (unsigned int iBin = 1; iBin < nBins; ++iBin)
-    {
-        // Get the number of event in the nominal universe and the universe supplied
-        const auto nEventsNom = m_allEventsNuEnergyNomTrue->GetBinContent(iBin);
-        const auto nEventUni = nuEnergyUniverse->GetBinContent(iBin);
-
-        // Get the ratio of the number of events - if not possible use a unit weight
-        const auto weight = (nEventsNom <= std::numeric_limits<float>::epsilon()) ? 1.f : (nEventUni / nEventsNom);
-
-        // Reweight the flux by this weight
-        const auto fluxNom = m_inputData.m_flux->GetBinContent(iBin);
-        const auto fluxUni = fluxNom * weight;
-
-        reweightedFlux->SetBinContent(iBin, fluxUni);
-    }
-
-    return reweightedFlux;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::map< std::string, std::shared_ptr<TH2F> > CrossSectionHelper::CrossSection::GetFluxVariations() const
-{
-    std::map< std::string, std::shared_ptr<TH2F> > outputMap;
-
-    for (const auto &param : m_inputData.m_fluxParameters)
-        outputMap.emplace(param, this->GetFluxVariations(param));
-
-    return outputMap;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH2F> CrossSectionHelper::CrossSection::GetFluxVariations(const std::string &systParameter) const
-{
-    if (std::find(m_inputData.m_fluxParameters.begin(), m_inputData.m_fluxParameters.end(), systParameter) == m_inputData.m_fluxParameters.end())
-        throw std::invalid_argument("CrossSection::GetFluxVariations - Input name \"" + systParameter + "\" is not one of the flux parameters supplied");
-
-    const auto &allEventsNuEnergyTrueVect = m_allEventsNuEnergyTrue.at(systParameter);
-    const auto nUniverses = allEventsNuEnergyTrueVect.size();
-
-    // Get the reweighted flux in each universe
-    float maxFlux = -std::numeric_limits<float>::max();
-    float minFlux = +std::numeric_limits<float>::max();
-    std::vector< std::shared_ptr<TH1F> > fluxVariations;
-    for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
-    {
-        const auto flux = this->GetReweightedFlux(allEventsNuEnergyTrueVect.at(iUni));
-
-        maxFlux = std::max(maxFlux, static_cast<float>(flux->GetMaximum()));
-        minFlux = std::min(minFlux, static_cast<float>(flux->GetMinimum()));
-        fluxVariations.push_back(flux);
-    }
-
-    // Setup the output histogram
-    const std::string name = "xSecHist_" + std::to_string(m_histCount++);
-
-    // Get the binning of the flux histogram
-    const unsigned int nBinsX = m_inputData.m_flux->GetNbinsX();
-    std::vector<float> edgesX;
-    for (unsigned int iBin = 1; iBin <= nBinsX; ++iBin)
-        edgesX.push_back(m_inputData.m_flux->GetBinLowEdge(iBin));
-
-    // Add the last bin edge
-    edgesX.push_back(m_inputData.m_flux->GetBinLowEdge(nBinsX) + m_inputData.m_flux->GetBinWidth(nBinsX));
-
-    // Get the binning in Y
-    const auto nBinsY = 100u;
-    std::vector<float> edgesY;
-    for (unsigned int iBin = 0; iBin < nBinsY; ++iBin)
-        edgesY.push_back(minFlux + static_cast<float>(iBin) * (maxFlux - minFlux) / static_cast<float>(nBinsY - 1));
-
-    auto outputHist = std::make_shared<TH2F>(name.c_str(), "", edgesX.size() - 1, edgesX.data(), edgesY.size() - 1, edgesY.data());
-
-    // Fill the histogram with the flux variations
-    for (const auto &flux : fluxVariations)
-    {
-        for (unsigned int iBinX = 1; iBinX <= nBinsX; ++iBinX)
-        {
-            const auto &fluxValue = flux->GetBinContent(iBinX);
-            const auto nuEnergy = 0.5f * (edgesX.at(iBinX - 1) + edgesX.at(iBinX));
-
-            outputHist->Fill(nuEnergy, fluxValue);
-        }
-    }
-
-    return outputHist;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::map< std::string, CrossSectionHelper::CovarianceBiasPair > CrossSectionHelper::CrossSection::GetCrossSectionCovarianceMatrices()
-{
-    std::map< std::string, CovarianceBiasPair> outputMap;
-
-    // Get the covariance matrices and bias vectors for the detector parameters
-    for (const auto &[runId, param] : m_inputData.m_detVarParameters)
-    {
-        const auto pair = this->GetCrossSectionCovarianceMatrixDetVar(runId, param);
-        outputMap.emplace(param, pair);
-    }
-
-    // Get the covariance matrices and bias vectors for the event-weighted systematic parameters
-    for (const auto &[param, nUniverses] : m_inputData.m_systUniverseSizesMap)
-    {
-        const auto pair = this->GetCrossSectionCovarianceMatrix(param);
-        outputMap.emplace(param, pair);
-    }
-
-    return outputMap;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::map< std::string, CrossSectionHelper::CovarianceBiasPair > CrossSectionHelper::CrossSection::GetSmearingMatrixCovarianceMatrices()
-{
-    std::map< std::string, CovarianceBiasPair> outputMap;
-
-    for (const auto &[param, nUniverses] : m_inputData.m_systUniverseSizesMap)
-    {
-        const auto pair = this->GetSmearingMatrixCovarianceMatrix(param);
-        outputMap.emplace(param, pair);
-    }
-
-    return outputMap;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::vector<float> CrossSectionHelper::GetSystematicWeights(const Event::Truth &truth, const std::string &parameter)
-{
-    if (!truth.systParamNames.IsSet() || !truth.systParamFirstValueIndex.IsSet() || !truth.systParamValues.IsSet())
-        throw std::invalid_argument("CrossSectionHelper::GetSystematicWeights - Systematic parameters are not set in the input truth information");
-
-    const auto &systParamNames = truth.systParamNames();
-    const auto &systParamFirstValueIndex = truth.systParamFirstValueIndex();
-    const auto &systParamValues = truth.systParamValues();
-
-    // Get the total number of parameters available
-    if (systParamNames.size() != systParamFirstValueIndex.size())
-        throw std::logic_error("CrossSectionHelper::GetSystematicWeights - Number of parameters is inconsistent in input truth information");
-
-    const auto nParameters = systParamNames.size();
-
-    // Find the requested parameter
-    const auto iter = std::find(systParamNames.begin(), systParamNames.end(), parameter);
-    if (iter == systParamNames.end())
-        throw std::invalid_argument("CrossSectionHelper::GetSystematicWeights - Unknown parameter \"" + parameter + "\"");
-
-    const unsigned int index = std::distance(systParamNames.begin(), iter);
-
-    // Get the first and last index in the weights vector
-    const unsigned int firstValueIndex = systParamFirstValueIndex.at(index);
-    const unsigned int lastValueIndex = ((index == nParameters - 1u) ? systParamValues.size() : systParamFirstValueIndex.at(index + 1u));
-
-    if (firstValueIndex >= systParamValues.size() || lastValueIndex >= systParamValues.size())
-        throw std::logic_error("CrossSectionHelper::GetSystematicWeights - Input systematic parameter first index vector is malformed");
-
-    // Pick out the weights corresponding to the requested paramter
-    std::vector<float> weights(std::next(systParamValues.begin(), firstValueIndex), std::next(systParamValues.begin(), lastValueIndex));
-
-    // Ensure the the input weights are valid
-    for (auto &weight : weights)
-    {
-        // If a weight is non-negative and non-infinite then it's fine!
-        if (weight >= 0.f && std::abs(std::abs(weight) - std::numeric_limits<float>::max()) > std::numeric_limits<float>::epsilon() && std::isfinite(weight))
-            continue;
-
-        // Force negative weights to be zero and infinite weights to be unity
-        weight = (weight < 0.f) ? 0.f : 1.f;
-    }
-
-    return weights;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::AddSystematicWeights(const Event::Truth &truth, const SystematicParamUniversesPairVector &params, SystematicWeightsMap &systWeightsMap)
-{
-    for (const auto &[name, nUniverses] : params)
-    {
-        const auto &weights = CrossSectionHelper::GetSystematicWeights(truth, name);
-
-        if (weights.size() != nUniverses)
-            throw std::logic_error("CrossSectionHelper::GetSystematicWeightsMap - Unexpected number of universes");
-
-        if (!systWeightsMap.emplace(name, weights).second)
-            throw std::invalid_argument("CrossSectionHelper::GetSystematicWeightsMap - Repeated systematic parameter: \"" + name + "\"");
-    }
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::AddMutuallyExclusiveSystematicWeights(const Event::Truth &truth, const MutuallyExclusiveParamVector &params, SystematicWeightsMap &systWeightsMap)
-{
-    for (const auto &[combinedName, nUniverses, paramsToCombine] : params)
-    {
-        // Build a SystematicParamUniversesPairVector for the parameters to combine
-        SystematicParamUniversesPairVector paramUniverseVector;
-        for (const auto &name : paramsToCombine)
-            paramUniverseVector.emplace_back(name, nUniverses);
-
-        // Get the weights for these parameters
-        SystematicWeightsMap weightsMap;
-        CrossSectionHelper::AddSystematicWeights(truth, paramUniverseVector, weightsMap);
-
-        // Build the combined weights
-        std::vector<float> combinedWeights;
-        for (unsigned int iUni = 0u; iUni < nUniverses; ++iUni)
-        {
-            // Get the weights in this universe from each parameter that aren't exactly 1.f
-            std::vector<float> variedWeights;
-            std::vector<std::string> variedParams;
-            for (const auto &name : paramsToCombine)
-            {
-                const auto &weight = weightsMap.at(name).at(iUni);
-
-                if (std::abs(weight - 1.f) > std::numeric_limits<float>::epsilon())
-                {
-                    variedWeights.push_back(weight);
-                    variedParams.push_back(name);
-                }
-            }
-
-            // Check the parameters are truly mutually exclusive
-            if (variedWeights.size() > 1)
-            {
-                std::string variedParamString = "";
-                for (const auto &name : variedParams)
-                    variedParamString += " " + name;
-
-                throw std::logic_error("CrossSectionHelper::AddMutuallyExclusiveSystematicWeights - Input params are not mutually exclusive. \
-                        Found a universe in which" + variedParamString + " each had values not equal to 1.f");
-            }
-
-            // Store the weight that was varied (or 1.f if there was no variation)
-            combinedWeights.push_back(variedWeights.empty() ? 1.f : variedWeights.front());
-        }
-
-        // Add the combined weights to the output map
-        if (!systWeightsMap.emplace(combinedName, combinedWeights).second)
-            throw std::invalid_argument("CrossSectionHelper::AddMutuallyExclusiveSystematicWeights - Repeated systematic parameter: \"" + combinedName + "\"");
-    }
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::AddBootstrapWeights(const unsigned int nBootstrapUniverses, SystematicWeightsMap &systWeightsMap)
-{
-    // Add the entry in the input map for the bootstrap systematic parameter - intially with unit weights
-    auto [iter, success] = systWeightsMap.emplace("bootstrap", std::vector<float>(nBootstrapUniverses, 1.f));
-
-    if (!success)
-        throw std::invalid_argument("CrossSectionHelper::AddBootstrapWeights - Bootstrap weights already exist in input map");
-
-    // Populate the MC stats systematic weights using the Poisson bootstrap method
-    std::poisson_distribution<int> poisson(1.f);
-    auto &weights = iter->second;
-    for (unsigned int iUni = 0; iUni < nBootstrapUniverses; ++iUni)
-        weights.at(iUni) = static_cast<float>(poisson(m_generator));
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::AddUnitWeights(const SystematicParamUniversesPairVector &params, SystematicWeightsMap &systWeightsMap)
-{
-    for (const auto &[name, nUniverses] : params)
-    {
-        if (!systWeightsMap.emplace(name, std::vector<float>(nUniverses, 1.f)).second)
-            throw std::logic_error("CrossSectionHelper::AddUnitWeights - Parameter \"" + name + "\" is already in the input map");
-    }
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::AddUnitWeights(const MutuallyExclusiveParamVector &params, SystematicWeightsMap &systWeightsMap)
-{
-    for (const auto &[name, nUniverses, exclusiveNames] : params)
-    {
-        if (!systWeightsMap.emplace(name, std::vector<float>(nUniverses, 1.f)).second)
-            throw std::logic_error("CrossSectionHelper::AddUnitWeights - Parameter \"" + name + "\" is already in the input map");
-    }
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-bool CrossSectionHelper::CheckSystematicWeightsMapDimensions(const SystematicWeightsMap &systWeightsMap, const SystematicUniverseSizesMap &systUniverseSizesMap)
-{
-    if (systWeightsMap.size() != systUniverseSizesMap.size())
-        return false;
-
-    for (const auto &[name, nUniverses] : systUniverseSizesMap)
-    {
-        const auto iter = systWeightsMap.find(name);
-        if (iter == systWeightsMap.end())
-            return false;
-
-        if (iter->second.size() != nUniverses)
-            return false;
-    }
-
-    return true;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-CrossSectionHelper::SystematicUniverseSizesMap CrossSectionHelper::GetSystematicUniverseSizesMap(const SystematicWeightsMap &systWeightsMap)
-{
-    // Sort the parameter names for reproducibility
-    std::vector<std::string> names;
-    for (const auto &[name, weights] : systWeightsMap)
-        names.push_back(name);
-
-    std::sort(names.begin(), names.end());
-
-    // Store the number of universes for each parameter
-    SystematicUniverseSizesMap outputMap;
-    for (const auto &name : names)
-    {
-        const auto nUniverses = systWeightsMap.at(name).size();
-        outputMap.emplace(name, nUniverses);
-    }
-
-    return outputMap;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-void CrossSectionHelper::GetExtendedBinEdges(const float min, const float max, const std::vector<float> &binEdges, std::vector<float> &extendedBinEdges, bool &hasUnderflow, bool &hasOverflow)
-{
-    if (!extendedBinEdges.empty())
-        throw std::invalid_argument("CrossSectionHelper::GetExtendedBinEdges - vector supplied to accept the output isn't empty");
-
-    // Ensure the bin edges are sorted
-    std::vector<float> sortedBinEdges(binEdges);
-    std::sort(sortedBinEdges.begin(), sortedBinEdges.end());
-
-    // Determine if underflow/overflow bins are required
-    const auto firstBinEdge = binEdges.front();
-    const auto lastBinEdge = binEdges.back();
-
-    if (firstBinEdge < min)
-        throw std::invalid_argument("CrossSectionHelper::GetExtendedBinEdges - lowest bin edge is less than the supplied minimum value");
-
-    if (lastBinEdge > max)
-        throw std::invalid_argument("CrossSectionHelper::GetExtendedBinEdges - highest bin edge is greater than the supplied maximum value");
-
-    hasUnderflow = (firstBinEdge - min > std::numeric_limits<float>::epsilon());
-    hasOverflow = (max - lastBinEdge > std::numeric_limits<float>::epsilon());
-
-    // Extend the bins as required
-    if (hasUnderflow)
-        extendedBinEdges.push_back(min);
-
-    extendedBinEdges.insert(extendedBinEdges.end(), binEdges.begin(), binEdges.end());
-
-    if (hasOverflow)
-        extendedBinEdges.push_back(max);
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-std::shared_ptr<TH1F> CrossSectionHelper::GetFluxHistogram(const Config::Flux &flux)
-{
-    auto pHist = std::make_shared<TH1F>(("fluxHist_" + std::to_string(m_histCount++)).c_str(), "", flux.binEdges.size() - 1, flux.binEdges.data());
-    const unsigned int nBins = pHist->GetNbinsX();
-
-    if (flux.energyBins.size() != nBins)
-        throw std::invalid_argument("CrossSectionHelper::GetFluxHistogram - Input number of input flux data points doesn't match the number of bins");
-
-    for (unsigned int iBin = 1; iBin <= nBins; ++iBin)
-    {
-        // ATTN ROOT enumerates bins from 1, but we are reading data from a vector that enumerates from 0
-        pHist->SetBinContent(iBin, flux.energyBins.at(iBin - 1));
-    }
-
-    return pHist;
-}
-
-// -----------------------------------------------------------------------------------------------------------------------------------------
-
-float CrossSectionHelper::GetTotalFlux(const std::shared_ptr<TH1F> &fluxHist)
+float CrossSectionHelper::FluxReweightor::GetIntegratedFlux(const std::shared_ptr<TH1F> &pFlux) const
 {
     float total = 0.f;
 
-    const unsigned int nBins = fluxHist->GetNbinsX();
-    for (unsigned int iBin = 1; iBin <= nBins; ++iBin)
+    for (unsigned int iBin = 1; iBin <= static_cast<unsigned int>(pFlux->GetNbinsX()); ++iBin)
     {
-        total += fluxHist->GetBinContent(iBin);
+        // ATTN the input flux has already been scaled by bin width
+        total += pFlux->GetBinContent(iBin);
     }
 
     return total;
@@ -1501,119 +119,1112 @@ float CrossSectionHelper::GetTotalFlux(const std::shared_ptr<TH1F> &fluxHist)
 
 // -----------------------------------------------------------------------------------------------------------------------------------------
 
-std::pair<TVector2, TVector2> CrossSectionHelper::GetUncertaintyEigenVectors(const CovarianceBiasPair &covarianceBias, const unsigned int iBin, const unsigned int jBin)
+float CrossSectionHelper::FluxReweightor::GetIntegratedNominalFlux() const
 {
-    // Pick out the covariance matrix
-    const auto &covarianceMatrix = covarianceBias.first;
-
-    // Make sure we have square matrices of the same dimensions
-    const auto &nBins = covarianceMatrix->GetNbinsX();
-    if (covarianceMatrix->GetNbinsY() != nBins)
-        throw std::invalid_argument("CrossSectionHelper::GetUncertaintyEigenVectors - Input covariance matrix is not square");
-
-    // Make sure the requested bins are valid
-    if (iBin == 0 || iBin > static_cast<unsigned int>(nBins))
-    {
-        throw std::range_error("CrossSectionHelper::GetUncertaintyEigenVectors - Supplied bin index iBin = " + std::to_string(iBin) +
-            " is out of range of input matrices: 1 - " + std::to_string(nBins));
-    }
-
-    if (jBin == 0 || jBin > static_cast<unsigned int>(nBins))
-    {
-        throw std::range_error("CrossSectionHelper::GetUncertaintyEigenVectors - Supplied bin index jBin = " + std::to_string(jBin) +
-            " is out of range of input matrices: 1 - " + std::to_string(nBins));
-    }
-
-    // Get the total error matrix elements for the bins we care about
-    // Here we take the sum of the covariance and bias matrices as the total error
-    const auto ii = covarianceMatrix->GetBinContent(iBin, iBin);
-    const auto ij = covarianceMatrix->GetBinContent(iBin, jBin);
-    const auto ji = covarianceMatrix->GetBinContent(jBin, iBin);
-    const auto jj = covarianceMatrix->GetBinContent(jBin, jBin);
-
-    // Get the discriminant of the eigen equation for this 2D covariance matrix
-    const auto trace = ii + jj;
-    const auto det = ii*jj - ij*ji;
-    const auto disc = trace*trace - 4*det;
-
-    if (disc < 0.f)
-        throw std::logic_error("CrossSectionHelper::GetUncertaintyEigenVectors - Eigen equation has no solutions");
-
-    // Get the eigenvalues
-    const auto rootDisc = std::pow(disc, 0.5f);
-    const auto lambdaMax = std::max(trace + rootDisc, trace - rootDisc) / 2.;
-    const auto lambdaMin = std::min(trace + rootDisc, trace - rootDisc) / 2.;
-
-    // Get the unit eigenvectors
-    const auto vectMaxUnit = TVector2(ij, lambdaMax - ii).Unit();
-    const auto vectMinUnit = TVector2(ij, lambdaMin - ii).Unit();
-
-    // Get the eigen vectors which have length of the root of the eigen values
-    const auto sigmaMax = std::pow(lambdaMax, 0.5f);
-    const auto sigmaMin = std::pow(lambdaMin, 0.5f);
-
-    const auto vectMax = sigmaMax * vectMaxUnit;
-    const auto vectMin = sigmaMin * vectMinUnit;
-
-    return std::pair<TVector2, TVector2>(vectMax, vectMin);
+    return this->GetIntegratedFlux(m_pFluxNominal);
 }
 
 // -----------------------------------------------------------------------------------------------------------------------------------------
 
-std::shared_ptr<TH2F> CrossSectionHelper::GetTotalCovarianceMatrix(const std::map< std::string, CovarianceBiasPair > &covarianceBiasPairs)
+float CrossSectionHelper::FluxReweightor::GetIntegratedFluxVariation(const std::string &paramName, const unsigned int universeIndex)
 {
-    std::shared_ptr<TH2F> totalCovarianceMatrix;
+    // Call the cache-function to either calculate the return value or retrieve it from the cache
+    return m_getIntegratedFluxVariationFunction(paramName, universeIndex);
+}
 
-    for (const auto &[paramName, covarianceBiasPair] : covarianceBiasPairs)
+// -----------------------------------------------------------------------------------------------------------------------------------------
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::CrossSection::CrossSection(const SystParams &systParams, const std::vector<float> &binEdges, const bool hasUnderflow, const bool hasOverflow, const bool scaleByBinWidth) :
+    m_systParams(systParams),
+    m_binEdges(binEdges),
+    m_metadata(binEdges.size() - 1, hasUnderflow, hasOverflow),
+    m_scaleByBinWidth(scaleByBinWidth),
+    m_pSignal_true_nom(CrossSectionHelper::GetTH1F(binEdges)),
+    m_pSignal_selected_recoTrue_nom(CrossSectionHelper::GetTH2F(binEdges)),
+    m_pBackground_selected_reco_nom(CrossSectionHelper::GetTH1F(binEdges)),
+    m_pBNBData_selected_reco(CrossSectionHelper::GetTH1F(binEdges))
+{
+    // Make sure the bin edges are valid
+    const int nAnalysisBins = (binEdges.size() - 1) - (hasUnderflow ? 1 : 0) - (hasOverflow ? 1 : 0);
+    if (nAnalysisBins <= 0)
+        throw std::invalid_argument("CrossSection::CrossSection - Insufficient bin edges provided!");
+
+    // Setup a dimensions map for the miscellaneous ("misc") multisim parameters.
+    // These are added in ad-hoc (i.e. they don't come from the weights stored in the truth information)
+    const SystDimensionsMap miscDimensions = {
+
+        // Here we apply a weight to each event using the Poisson bootstrap method to estimate the MC stats uncertainty
+        { "bootstrap", systParams.nBootstrapUniverses },
+
+        // Here we apply a 2-universe multisim in which we weight the dirt up and down by 100%
+        { "dirt", 2 }
+    };
+
+    // Setup the maps to hold the histograms for the multisim parameters
+    m_signal_true_multisims.emplace("flux", CrossSectionHelper::GetSystTH1FMap(binEdges, systParams.fluxDimensions));
+    m_signal_true_multisims.emplace("xsec", CrossSectionHelper::GetSystTH1FMap(binEdges, systParams.xsecDimensions));
+    m_signal_true_multisims.emplace("misc", CrossSectionHelper::GetSystTH1FMap(binEdges, miscDimensions));
+
+    m_signal_selected_recoTrue_multisims.emplace("flux", CrossSectionHelper::GetSystTH2FMap(binEdges, systParams.fluxDimensions));
+    m_signal_selected_recoTrue_multisims.emplace("xsec", CrossSectionHelper::GetSystTH2FMap(binEdges, systParams.xsecDimensions));
+    m_signal_selected_recoTrue_multisims.emplace("misc", CrossSectionHelper::GetSystTH2FMap(binEdges, miscDimensions));
+
+    m_background_selected_reco_multisims.emplace("flux", CrossSectionHelper::GetSystTH1FMap(binEdges, systParams.fluxDimensions));
+    m_background_selected_reco_multisims.emplace("xsec", CrossSectionHelper::GetSystTH1FMap(binEdges, systParams.xsecDimensions));
+    m_background_selected_reco_multisims.emplace("misc", CrossSectionHelper::GetSystTH1FMap(binEdges, miscDimensions));
+
+    // Setup the maps to hold the histograms for the unisim parameters
+    m_signal_true_unisims.emplace("detector", CrossSectionHelper::GetSystUnisimTH1FMap(binEdges, systParams.detVarDimensions));
+    m_signal_selected_recoTrue_unisims.emplace("detector", CrossSectionHelper::GetSystUnisimTH2FMap(binEdges, systParams.detVarDimensions));
+    m_background_selected_reco_unisims.emplace("detector", CrossSectionHelper::GetSystUnisimTH1FMap(binEdges, systParams.detVarDimensions));
+
+    // ATTN we could in theory include other unisim parameters here using a different name. At the moment we only use "detector".
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+void CrossSectionHelper::CrossSection::AddSignalEvent(const float recoValue, const float trueValue, const bool isSelected, const float nominalWeight, const SystFloatMap &fluxWeights, const SystFloatMap &xsecWeights)
+{
+    // Get the miscellaneous weights
+    const SystFloatMap miscWeights = {
+
+        // Generate the bootstrap weights
+        { "bootstrap", CrossSectionHelper::GenerateBootstrapWeights(m_systParams.nBootstrapUniverses) },
+
+        // A signal event is never dirt - so just use a unit weight
+        { "dirt", {1.f, 1.f} }
+    };
+
+    m_pSignal_true_nom->Fill(trueValue, nominalWeight);
+    CrossSectionHelper::FillSystTH1FMap(trueValue, nominalWeight, fluxWeights, m_signal_true_multisims.at("flux"));
+    CrossSectionHelper::FillSystTH1FMap(trueValue, nominalWeight, xsecWeights, m_signal_true_multisims.at("xsec"));
+    CrossSectionHelper::FillSystTH1FMap(trueValue, nominalWeight, miscWeights, m_signal_true_multisims.at("misc"));
+
+    if (!isSelected)
+        return;
+
+    m_pSignal_selected_recoTrue_nom->Fill(recoValue, trueValue, nominalWeight);
+    CrossSectionHelper::FillSystTH2FMap(recoValue, trueValue, nominalWeight, fluxWeights, m_signal_selected_recoTrue_multisims.at("flux"));
+    CrossSectionHelper::FillSystTH2FMap(recoValue, trueValue, nominalWeight, xsecWeights, m_signal_selected_recoTrue_multisims.at("xsec"));
+    CrossSectionHelper::FillSystTH2FMap(recoValue, trueValue, nominalWeight, miscWeights, m_signal_selected_recoTrue_multisims.at("misc"));
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+void CrossSectionHelper::CrossSection::AddSignalEventDetVar(const float recoValue, const float trueValue, const bool isSelected, const float nominalWeight, const std::string &paramName)
+{
+    if (!static_cast<bool>(m_signal_true_unisims.at("detector").count(paramName)))
+        throw std::invalid_argument("CrossSection::AddSignalEventDetVar - Unknown parameter or CV sample name \"" + paramName + "\"");
+
+    m_signal_true_unisims.at("detector").at(paramName)->Fill(trueValue, nominalWeight);
+
+    if (!isSelected)
+        return;
+
+    m_signal_selected_recoTrue_unisims.at("detector").at(paramName)->Fill(recoValue, trueValue, nominalWeight);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+void CrossSectionHelper::CrossSection::AddSelectedBackgroundEvent(const float recoValue, const bool isDirt, const float nominalWeight, const SystFloatMap &fluxWeights, const SystFloatMap &xsecWeights)
+{
+    // The weight to apply as a result of dirt
+    //   - For non-dirt events, just use zero (i.e. no change from the nominal simulation)
+    //   - For dirt event we here use +-100%
+    const auto dirtWeightDelta = isDirt ? 1.f : 0.f;
+
+    // Get the miscellaneous weights
+    const SystFloatMap miscWeights = {
+
+        // Generate the bootstrap weights
+        { "bootstrap", CrossSectionHelper::GenerateBootstrapWeights(m_systParams.nBootstrapUniverses) },
+
+        // Apply the dirt weight
+        { "dirt", {1.f - dirtWeightDelta, 1.f + dirtWeightDelta} }
+    };
+
+    m_pBackground_selected_reco_nom->Fill(recoValue, nominalWeight);
+    CrossSectionHelper::FillSystTH1FMap(recoValue, nominalWeight, fluxWeights, m_background_selected_reco_multisims.at("flux"));
+    CrossSectionHelper::FillSystTH1FMap(recoValue, nominalWeight, xsecWeights, m_background_selected_reco_multisims.at("xsec"));
+    CrossSectionHelper::FillSystTH1FMap(recoValue, nominalWeight, miscWeights, m_background_selected_reco_multisims.at("misc"));
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+void CrossSectionHelper::CrossSection::AddSelectedBackgroundEventDetVar(const float recoValue, const float nominalWeight, const std::string &paramName)
+{
+    if (!static_cast<bool>(m_background_selected_reco_unisims.at("detector").count(paramName)))
+        throw std::invalid_argument("CrossSection::AddSelectedBackgroundEventDetVar - Unknown parameter or CV sample name \"" + paramName + "\"");
+
+    m_background_selected_reco_unisims.at("detector").at(paramName)->Fill(recoValue, nominalWeight);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+void CrossSectionHelper::CrossSection::AddSelectedBNBDataEvent(const float recoValue)
+{
+    m_pBNBData_selected_reco->Fill(recoValue);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::CrossSection::GetBinWidths() const
+{
+    if (!m_scaleByBinWidth)
     {
-        const auto &covarianceMatrix = covarianceBiasPair.first;
-        const auto &biasVector = covarianceBiasPair.second;
+        // If we shouldn't scale by bin width, then just return 1.f for the width of each bin
+        return ubsmear::UBMatrix(std::vector<float>(m_metadata.GetNBins(), 1.f), m_metadata.GetNBins(), 1);
+    }
 
-        // Check the binning is sensisble
-        const unsigned int nBins = covarianceMatrix->GetNbinsX();
-        if (static_cast<unsigned int>(covarianceMatrix->GetNbinsY()) != nBins)
-            throw std::invalid_argument("CrossSectionHelper::GetTotalCovarianceMatrix - Input covariance matrix is not square");
+    std::vector<float> elements;
+    for (unsigned int iBin = 1; iBin < m_binEdges.size(); ++iBin)
+    {
+        const auto lowerEdge = m_binEdges.at(iBin - 1);
+        const auto upperEdge = m_binEdges.at(iBin);
+        const auto width = upperEdge - lowerEdge;
 
-        if (static_cast<unsigned int>(biasVector->GetNbinsX()) != nBins)
-            throw std::invalid_argument("CrossSectionHelper::GetTotalCovarianceMatrix - Input bias vector has different dimensions to covariance matrix");
+        if (width <= std::numeric_limits<float>::epsilon())
+            throw std::logic_error("CrossSection::GetBinWidthVector - Found a bin with zero or negative width");
 
-        // If this is our first entry, then setup the binning of the output covariance matrix
-        if (!totalCovarianceMatrix)
+        elements.push_back(width);
+    }
+
+    return ubsmear::UBMatrix(elements, m_metadata.GetNBins(), 1);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::CrossSection::GetCrossSection(const ubsmear::UBMatrix &selected, const ubsmear::UBMatrix &backgrounds, const float integratedFlux, const float exposurePOT, const float nTargets) const
+{
+    // Check the input event rate has the right dimensions
+    if (selected.GetColumns() != 1)
+        throw std::invalid_argument("CrossSection::GetCrossSection - Input event rate isn't a column vector");
+
+    if (selected.GetRows() != m_metadata.GetNBins())
+        throw std::invalid_argument("CrossSection::GetCrossSection - Input event rate has the wrong number of bins");
+
+    // Check the input backgrounds have the right dimensions
+    if (backgrounds.GetColumns() != 1)
+        throw std::invalid_argument("CrossSection::GetCrossSection - Input background rate isn't a column vector");
+
+    if (backgrounds.GetRows() != m_metadata.GetNBins())
+        throw std::invalid_argument("CrossSection::GetCrossSection - Input background rate has the wrong number of bins");
+
+    // Get the normalisation factor
+    const auto norm = integratedFlux * exposurePOT * nTargets;
+    if (norm <= std::numeric_limits<float>::epsilon())
+        throw std::invalid_argument("CrossSection::GetCrossSection - Product of flux, exposure and targets non-positive");
+
+    // Get the column vector of bin widths
+    const auto binWidths = this->GetBinWidths();
+
+    // Now get the flux-integrated, forward-folded cross-section by
+    //   - Subtracting the backgrounds
+    //   - Scaling by the normalisation factor
+    //   - Scaling (element wise) by the bin-widths
+    return ubsmear::ElementWiseOperation(
+        (selected - backgrounds),
+        binWidths * norm,
+        [](const float numerator, const float denominator) { return numerator / denominator; }
+    );
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+std::shared_ptr<ubsmear::UBMatrix> CrossSectionHelper::CrossSection::GetSmearingMatrix(const std::shared_ptr<TH1F> &pSignal_true, const std::shared_ptr<TH2F> &pSignal_selected_recoTrue) const
+{
+    // Get the input histograms as a matrix
+    const auto allTrue = CrossSectionHelper::GetMatrixFromHist(pSignal_true);
+    const auto selectedRecoTrue = CrossSectionHelper::GetMatrixFromHist(pSignal_selected_recoTrue);
+
+    // Check theh matrices have the expected sizes
+    if (!ubsmear::UBMatrixHelper::IsSquare(selectedRecoTrue))
+        throw std::invalid_argument("CrossSection::GetSmearingMatrix - input reco-true matrix isn't square");
+
+    const auto nBins = allTrue.GetRows();
+    if (selectedRecoTrue.GetRows() != nBins)
+        throw std::invalid_argument("CrossSection::GetSmearingMatrix - the input objects differ in number of bins");
+
+    // Get the elements of the smearing matrix
+    std::vector<float> elements;
+    for (unsigned int iReco = 0; iReco < nBins; ++iReco)
+    {
+        for (unsigned int iTrue = 0; iTrue < nBins; ++iTrue)
         {
-            totalCovarianceMatrix = std::make_shared<TH2F>(("totalCovariance_" + std::to_string(m_histCount++)).c_str(), "", nBins, 0, nBins, nBins, 0, nBins);
+            // Get the total number of signal events in this true bin
+            const auto denominator = allTrue.At(iTrue, 0);
 
-            // Initialize the total covariance matrix with zeros
-            for (unsigned int iBin = 1; iBin <= nBins; ++iBin)
-            {
-                for (unsigned int jBin = 1; jBin <= nBins; ++jBin)
-                {
-                    totalCovarianceMatrix->SetBinContent(iBin, jBin, 0.f);
-                }
-            }
+            // ATTN. If we find a bin in which there are no signal events, then there's no sensible value for the smearing matrix element
+            // Here we return null to indicate that the smearing matrix can't be found
+            if (denominator <= std::numeric_limits<float>::epsilon())
+                return std::shared_ptr<ubsmear::UBMatrix>(nullptr);
+
+            elements.push_back(selectedRecoTrue.At(iReco, iTrue) / denominator);
         }
-        // Otherwise make sure that we have the expected number of bins
-        else if (static_cast<unsigned int>(totalCovarianceMatrix->GetNbinsX()) != nBins)
+    }
+
+    return std::make_shared<ubsmear::UBMatrix>(elements, nBins, nBins);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::CrossSection::GetSignalSelectedTrue(const std::shared_ptr<TH2F> &pSignal_selected_recoTrue) const
+{
+    // Get the input histogram as a matrix
+    const auto recoTrueMatrix = CrossSectionHelper::GetMatrixFromHist(pSignal_selected_recoTrue);
+
+    // Integrate over the reco bins
+    std::vector<float> elements;
+    for (unsigned int iTrue = 0; iTrue < recoTrueMatrix.GetColumns(); ++iTrue)
+    {
+        float sum = 0.f;
+        for (unsigned int iReco = 0; iReco < recoTrueMatrix.GetRows(); ++iReco)
         {
-            throw std::invalid_argument("CrossSectionHelper::GetTotalCovarianceMatrix - Sizes of input covaraince matrices are not equal");
+            sum += recoTrueMatrix.At(iReco, iTrue);
         }
 
-        // Add to the total covariance matrix
-        for (unsigned int iBin = 1u; iBin <= nBins; ++iBin)
-        {
-            for (unsigned int jBin = 1u; jBin <= nBins; ++jBin)
-            {
-                const auto covariance = covarianceMatrix->GetBinContent(iBin, jBin);
-                const auto bias = biasVector->GetBinContent(iBin) * biasVector->GetBinContent(jBin);
+        elements.push_back(sum);
+    }
 
-                // Add the covariance and bias to the current total
-                const auto currentValue = totalCovarianceMatrix->GetBinContent(iBin, jBin);
-                const auto updatedValue = currentValue + covariance + bias;
-                totalCovarianceMatrix->SetBinContent(iBin, jBin, updatedValue);
+    // Return the result as a column vector
+    return ubsmear::UBMatrix(elements, recoTrueMatrix.GetColumns(), 1);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::CrossSection::GetBNBDataCrossSection(const ScalingData &scalingData) const
+{
+    // Get the number of selected events in BNB data
+    const auto selected = CrossSectionHelper::GetMatrixFromHist(m_pBNBData_selected_reco);
+
+    // Get the number of predicted backgrounds in the nominal universe
+    const auto backgrounds = CrossSectionHelper::GetMatrixFromHist(m_pBackground_selected_reco_nom);
+
+    // Get the integrated flux in the nominal universe
+    const auto integratedFlux = scalingData.fluxReweightor->GetIntegratedNominalFlux();
+
+    // Get the cross-section
+    return this->GetCrossSection(selected, backgrounds, integratedFlux, scalingData.exposurePOT, scalingData.nTargets);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::CrossSection::GetPredictedCrossSection(const ScalingData &scalingData) const
+{
+    // Get the number of signal events in the nominal simulation in true bins
+    const auto signal = CrossSectionHelper::GetMatrixFromHist(m_pSignal_true_nom);
+
+    // Get the number of backgrounds is zero as we are effectively applying a "perfect" selection
+    const auto zeroVector = ubsmear::UBMatrixHelper::GetZeroMatrix(signal.GetRows(), 1);
+
+    // Get the integrated flux in the nominal universe
+    const auto integratedFlux = scalingData.fluxReweightor->GetIntegratedNominalFlux();
+
+    // Get the cross-section
+    return this->GetCrossSection(signal, zeroVector, integratedFlux, scalingData.exposurePOT, scalingData.nTargets);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+std::shared_ptr<ubsmear::UBMatrix> CrossSectionHelper::CrossSection::GetSmearingMatrixNominal() const
+{
+    // Get the smearing matrix in the nominal universe
+    return this->GetSmearingMatrix(m_pSignal_true_nom, m_pSignal_selected_recoTrue_nom);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::CrossSection::GetSmearingMatrix() const
+{
+    // Get the smearing matrix in the nominal universe
+    const auto pSmearingMatrix = this->GetSmearingMatrixNominal();
+
+    // Check we were able to find the smearing matrix
+    if (!pSmearingMatrix)
+        throw std::logic_error("CrossSection::GetSmearingMatrix - There's a bin with no signal events! Can't get the smearing matrix");
+
+    return *pSmearingMatrix;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::CrossSection::GetBNBDataCrossSectionStatUncertainty(const ScalingData &scalingData) const
+{
+    // Get the number of selected events in BNB data
+    const auto selected = CrossSectionHelper::GetMatrixFromHist(m_pBNBData_selected_reco);
+
+    // Get a vector containing the count uncertainties (Poisson) on each bin
+    std::vector<float> elements;
+    const auto nBins = selected.GetRows();
+    for (unsigned int iBin = 0; iBin < nBins; ++iBin)
+    {
+        elements.push_back(AnalysisHelper::GetCountUncertainty(selected.At(iBin, 0)));
+    }
+
+    const ubsmear::UBMatrix selectedUncertainty(elements, nBins, 1);
+
+    // Get the integrated flux in the nominal universe
+    const auto integratedFlux = scalingData.fluxReweightor->GetIntegratedNominalFlux();
+
+    // Scale the uncertainty on the number of selected events by same factor as the cross-section.
+    // ATTN to apply the scaling we reuse the GetCrossSection function but intead pass it the count uncertainties as "selected" a zero
+    // vector as the "backgrounds"
+    const auto zeroVector = ubsmear::UBMatrixHelper::GetZeroMatrix(nBins, 1);
+    return this->GetCrossSection(selectedUncertainty, zeroVector, integratedFlux, scalingData.exposurePOT, scalingData.nTargets);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystBiasCovarianceMap CrossSectionHelper::CrossSection::GetBNBDataCrossSectionSystUncertainties(const ScalingData &scalingData) const
+{
+    // Setup the output map
+    SystBiasCovarianceMap outputMap;
+
+    // Add the multisim parameters
+    // ATTN the use of m_signal_true_multisims is arbitrary, we could use any multisims map
+    for (const auto &[group, map] : m_signal_true_multisims)
+    {
+        for (const auto &entry : map)
+        {
+            const auto &paramName = entry.first;
+            outputMap[group].emplace(paramName, this->GetBNBDataCrossSectionDistributionParams(group, paramName, scalingData));
+        }
+    }
+
+    // Add the unisim parameters
+    // ATTN again the use of m_signal_true_unisims is arbitrary, we could use any unisims map
+    for (const auto &[group, map] : m_signal_true_unisims)
+    
+    {
+        // ATTN for now we only have detector parameters that are unisims, if others are added later then we would need to add a clause here
+        // to get the relevent dimensions.
+        if (group != "detector")
+            throw std::logic_error("CrossSection::GetBNBDataCrossSectionSystUncertainties - Don't know dimensions of group: \"" + group + "\"");
+
+        const auto dimensions = m_systParams.detVarDimensions;
+
+        for (const auto &entry : map)
+        {
+            const auto &paramName = entry.first;
+            const auto &cvName = dimensions.at(paramName);
+
+            outputMap[group].emplace(paramName, this->GetBNBDataCrossSectionDistributionParamsUnisim(group, paramName, cvName, scalingData));
+        }
+    }
+
+    return outputMap;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystBiasCovarianceMap CrossSectionHelper::CrossSection::GetSmearingMatrixSystUncertainties() const
+{
+    // Setup the output map
+    SystBiasCovarianceMap outputMap;
+
+    // Add the multisim parameters
+    // ATTN the use of m_signal_true_multisims is arbitrary, we could use any multisims map
+    for (const auto &[group, map] : m_signal_true_multisims)
+    {
+        for (const auto &entry : map)
+        {
+            const auto &paramName = entry.first;
+            outputMap[group].emplace(paramName, this->GetSmearingMatrixDistributionParams(group, paramName));
+        }
+    }
+
+    // Add the unisim parameters
+    // ATTN again the use of m_signal_true_unisims is arbitrary, we could use any unisims map
+    for (const auto &[group, map] : m_signal_true_unisims)
+    {
+        // ATTN for now we only have detector parameters that are unisims, if others are added later then we would need to add a clause here
+        // to get the relevent dimensions.
+        if (group != "detector")
+            throw std::logic_error("CrossSection::GetSmearingMatrixSystUncertainties - Don't know dimensions of group: \"" + group + "\"");
+
+        const auto dimensions = m_systParams.detVarDimensions;
+
+        for (const auto &entry : map)
+        {
+            const auto &paramName = entry.first;
+            const auto &cvName = dimensions.at(paramName);
+
+            outputMap[group].emplace(paramName, this->GetSmearingMatrixDistributionParamsUnisim(group, paramName, cvName));
+        }
+    }
+
+    return outputMap;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystBiasCovariancePair CrossSectionHelper::CrossSection::GetPredictedCrossSectionStatUncertainty(const ScalingData &scalingData) const
+{
+    // Get the number of universes for this systematic parameter
+    // ATTN this choice of m_signal_true_multisims here is arbitrary, any of the mutlisims maps would do
+    const auto nUniverses = m_systParams.nBootstrapUniverses;
+
+    // Get the nominal cross-section
+    const auto pXSecNom = std::make_shared<ubsmear::UBMatrix>(this->GetPredictedCrossSection(scalingData));
+
+    // Get the number of backgrounds is zero as we are effectively applying a "perfect" selection
+    const auto zeroVector = ubsmear::UBMatrixHelper::GetZeroMatrix(pXSecNom->GetRows(), 1);
+
+    // Get the integrated flux in the nominal universe
+    const auto integratedFlux = scalingData.fluxReweightor->GetIntegratedNominalFlux();
+
+    // The bootstrap universes
+    const auto universes = m_signal_true_multisims.at("misc").at("bootstrap");
+
+    // Get the distirbution parameters
+    return this->GetDistributionParams([&](const unsigned int &iUni)
+    {
+        // This is the function we call to get the value in each universe
+
+        // Get the number of signal events in the nominal simulation in true bins
+        const auto signal = CrossSectionHelper::GetMatrixFromHist(universes.at(iUni));
+
+        // Get the cross-section
+        return std::make_shared<ubsmear::UBMatrix> ( this->GetCrossSection(signal, zeroVector, integratedFlux, scalingData.exposurePOT, scalingData.nTargets) );
+
+    }, nUniverses, pXSecNom);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::CrossSection::GetBNBDataCrossSectionInUniverse(const std::string &group, const std::string &paramName, const unsigned int universeIndex, const ScalingData &scalingData) const
+{
+    // ATTN for speed, this function (and other similar functions) doesn't explicitly check if the input group, paramName, universeIndex is
+    // valid. Instead we rely on the internal range checking of the .at() method for STL containers. This is for peformance reasons as this
+    // function gets called once per universe!
+
+    // Get the number of selected signal events in the supplied universe
+    const auto signalSelected = this->GetSignalSelectedTrue(m_signal_selected_recoTrue_multisims.at(group).at(paramName).at(universeIndex));
+
+    // Get the number of predicted backgrounds in the supplied universe
+    const auto backgrounds = CrossSectionHelper::GetMatrixFromHist(m_background_selected_reco_multisims.at(group).at(paramName).at(universeIndex));
+
+    // Get the total number of selected events
+    const auto selected = signalSelected + backgrounds;
+
+    // Get the integrated flux in the supplied universe (if it's not a flux parameter, then use the nominal universe)
+    const auto integratedFlux = (
+        group == "flux"
+            ? scalingData.fluxReweightor->GetIntegratedFluxVariation(paramName, universeIndex)
+            : scalingData.fluxReweightor->GetIntegratedNominalFlux()
+    );
+
+    // Get the cross-section
+    return this->GetCrossSection(selected, backgrounds, integratedFlux, scalingData.exposurePOT, scalingData.nTargets);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::CrossSection::GetBNBDataCrossSectionForUnisim(const std::string &group, const std::string &paramName, const ScalingData &scalingData) const
+{
+    // Get the number of selected signal events in the supplied universe
+    const auto signalSelected = this->GetSignalSelectedTrue(m_signal_selected_recoTrue_unisims.at(group).at(paramName));
+
+    // Get the number of predicted backgrounds in the supplied universe
+    const auto backgrounds = CrossSectionHelper::GetMatrixFromHist(m_background_selected_reco_unisims.at(group).at(paramName));
+
+    // Get the total number of selected events
+    const auto selected = signalSelected + backgrounds;
+
+    // Get the integrated flux in the nominal universe
+    const auto integratedFlux = scalingData.fluxReweightor->GetIntegratedNominalFlux();
+
+    // Get the cross-section
+    return this->GetCrossSection(selected, backgrounds, integratedFlux, scalingData.exposurePOT, scalingData.nTargets);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystBiasCovariancePair CrossSectionHelper::CrossSection::GetDistributionParams(const std::function<std::shared_ptr<ubsmear::UBMatrix>(const unsigned int)> &func, const unsigned int nUniverses, const std::shared_ptr<ubsmear::UBMatrix> &pNominal) const
+{
+    if (nUniverses == 0)
+        throw std::logic_error("CrossSection::GetDisributionParams - No universes supplied");
+
+    // Insist the nominal input is a column vector
+    if (pNominal->GetColumns() != 1)
+        throw std::logic_error("CrossSection::GetDisributionParams - Input nominal is not a column vector");
+
+    // Get the number of bins
+    const auto nBins = pNominal->GetRows();
+
+    // Setup some empty matrices to hold the mean vector and error matrices
+    auto meanSum = ubsmear::UBMatrixHelper::GetZeroMatrix(nBins, 1);
+    auto errorMatrixSum = ubsmear::UBMatrixHelper::GetZeroMatrix(nBins, nBins);
+
+    // Loop over the universes
+    unsigned int nValidUniverses = 0u;
+    for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
+    {
+        // Get the value in this universe
+        const auto pUniverse = func(iUni);
+
+        // Insist that we get a valid quanitiy in this universes
+        if (!pUniverse)
+            continue;
+
+        // Count the number of valid universes
+        nValidUniverses++;
+
+        // Loop over the bins
+        // ATTN for speed we don't expcitly check here if the universe has the same number of bins as the nominal. Instead we rely on the
+        // internal range checking of the ubsmear::UBMatrix class
+        for (unsigned int iBin = 0; iBin < nBins; ++iBin)
+        {
+            // Add up the cross-sections
+            meanSum.SetElement(iBin, 0, meanSum.At(iBin, 0) + pUniverse->At(iBin, 0));
+
+            // Loop over the bins again
+            // ATTN the error matrix is symmetic so here we only set one half of the off-diagonals within the universe loop and then copy
+            // them over to the other half afterwards. This is done for performance reasons as it halves the number of inserts required
+            const auto diffI = pUniverse->At(iBin, 0) - pNominal->At(iBin, 0);
+            for (unsigned int jBin = 0; jBin <= iBin; ++jBin)
+            {
+                const auto diffJ = pUniverse->At(jBin, 0) - pNominal->At(jBin, 0);
+
+                // Add up the error matrix elements
+                errorMatrixSum.SetElement(iBin, jBin, errorMatrixSum.At(iBin, jBin) + diffI*diffJ);
             }
         }
     }
 
-    return totalCovarianceMatrix;
+    // Scale the sums by the number of universes
+    if (nValidUniverses == 0)
+        throw std::logic_error("CrossSection::GetDisributionParams - Desired quantity was invalid in all universes");
+
+    const auto scaleFactor = 1.f / static_cast<float>(nValidUniverses);
+    const auto mean = meanSum * scaleFactor;
+    const auto errorMatrix = errorMatrixSum * scaleFactor;
+
+    // Get the bias vector
+    const auto pBias = std::make_shared<ubsmear::UBMatrix>(mean - *pNominal);
+
+    // Get the covariance matrix
+    auto pCovarianceMatrix = std::make_shared<ubsmear::UBMatrix>(ubsmear::UBMatrixHelper::GetZeroMatrix(nBins, nBins));
+    for (unsigned int iBin = 0; iBin < nBins; ++iBin)
+    {
+        for (unsigned int jBin = 0; jBin <= iBin; ++jBin)
+        {
+            // ATTN here we use the fact that V_ij = E_ij - b_i*b_j
+            //   - E_ij = error matrix element
+            //   - V_ij = covariance matrix element
+            //   - b_i = bias vector element
+            //
+            // This is done instead of calculating the covariance directly as this only requires one pass through the universe loop.
+            const auto covariance = errorMatrix.At(iBin, jBin) - pBias->At(iBin, 0)*pBias->At(jBin, 0);
+
+            pCovarianceMatrix->SetElement(iBin, jBin, covariance);
+            pCovarianceMatrix->SetElement(jBin, iBin, covariance);
+        }
+    }
+
+    return { pBias, pCovarianceMatrix };
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystBiasCovariancePair CrossSectionHelper::CrossSection::GetDistributionParamsUnisim(const std::shared_ptr<ubsmear::UBMatrix> &pVaried, const std::shared_ptr<ubsmear::UBMatrix> &pCentralValue, const std::shared_ptr<ubsmear::UBMatrix> &pNominal) const
+{
+    // Check the input matrices are valid
+    if (!pVaried || !pCentralValue || !pNominal)
+        throw std::invalid_argument("CrossSection::GetDistributionParamsUnisim - One or more of the inputs are not valid");
+
+    // Check ths input matrices have the desired dimensions
+    if (pVaried->GetColumns() != 1 || pCentralValue->GetColumns() != 1 || pNominal->GetColumns() != 1)
+        throw std::invalid_argument("CrossSection::GetDistributionParamsUnisim - One or more of the inputs are not a column vector");
+
+    const auto nBins = pVaried->GetRows();
+    if (pCentralValue->GetRows() != nBins || pNominal->GetRows() != nBins)
+        throw std::invalid_argument("CrossSection::GetDistributionParamsUnisim - Input column vectors have different sizes");
+
+    // Get the fractional bias of the cross-section away from the central value
+    const auto fracBias = ubsmear::ElementWiseOperation(
+        (*pVaried - *pCentralValue), *pCentralValue,
+        [](const float numerator, const float denominator)
+        {
+            if (std::abs(denominator) <= std::numeric_limits<float>::epsilon())
+                throw std::logic_error("CrossSection::GetDistributionParamsUnisim - Input nominal cross-section has a zero bin");
+
+            return numerator / denominator;
+        }
+    );
+
+    // Apply the fractional bias to the nominal cross-section
+    const auto bias = ubsmear::ElementWiseOperation(
+        fracBias, *pNominal,
+        [](const float lhs, const float rhs) { return lhs * rhs; }
+    );
+
+    // Return the result (use a zero matrix for the covariance)
+    return {
+        std::make_shared<ubsmear::UBMatrix>(bias),
+        std::make_shared<ubsmear::UBMatrix>(ubsmear::UBMatrixHelper::GetZeroMatrix(nBins, nBins))
+    };
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystBiasCovariancePair CrossSectionHelper::CrossSection::GetBNBDataCrossSectionDistributionParams(const std::string &group, const std::string &paramName, const ScalingData &scalingData) const
+{
+    // Get the number of universes for this systematic parameter
+    // ATTN this choice of m_signal_true_multisims here is arbitrary, any of the mutlisims maps would do
+    const auto nUniverses = m_signal_true_multisims.at(group).at(paramName).size();
+
+    // Get the nominal cross-section
+    const auto pXSecNom = std::make_shared<ubsmear::UBMatrix>(this->GetBNBDataCrossSection(scalingData));
+
+    // Get the distirbution parameters
+    return this->GetDistributionParams([&](const unsigned int &iUni)
+    {
+        // This is the function we call to get the value in each universe
+        return std::make_shared<ubsmear::UBMatrix>( this->GetBNBDataCrossSectionInUniverse(group, paramName, iUni, scalingData) );
+
+    }, nUniverses, pXSecNom);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystBiasCovariancePair CrossSectionHelper::CrossSection::GetBNBDataCrossSectionDistributionParamsUnisim(const std::string &group, const std::string &paramName, const std::string &cvName, const ScalingData &scalingData) const
+{
+    // Get the cross-section using the unisim sample, in the central value sample and in the nominal universe
+    const auto pXSec = std::make_shared<ubsmear::UBMatrix>( this->GetBNBDataCrossSectionForUnisim(group, paramName, scalingData) );
+    const auto pXSecCV = std::make_shared<ubsmear::UBMatrix>( this->GetBNBDataCrossSectionForUnisim(group, cvName, scalingData) );
+    const auto pXSecNom = std::make_shared<ubsmear::UBMatrix>( this->GetBNBDataCrossSection(scalingData) );
+
+    // Get the parameters
+    return this->GetDistributionParamsUnisim(pXSec, pXSecCV, pXSecNom);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+std::shared_ptr<ubsmear::UBMatrix> CrossSectionHelper::CrossSection::GetSmearingMatrixInUniverse(const std::string &group, const std::string &paramName, const unsigned int universeIndex) const
+{
+    return this->GetSmearingMatrix(m_signal_true_multisims.at(group).at(paramName).at(universeIndex), m_signal_selected_recoTrue_multisims.at(group).at(paramName).at(universeIndex));
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+std::shared_ptr<ubsmear::UBMatrix> CrossSectionHelper::CrossSection::GetSmearingMatrixForUnisim(const std::string &group, const std::string &paramName) const
+{
+    return this->GetSmearingMatrix(m_signal_true_unisims.at(group).at(paramName), m_signal_selected_recoTrue_unisims.at(group).at(paramName));
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystBiasCovariancePair CrossSectionHelper::CrossSection::GetSmearingMatrixDistributionParams(const std::string &group, const std::string &paramName) const
+{
+    // Get the number of universes for this systematic parameter
+    // ATTN this choice of m_signal_true_multisims here is arbitrary, any of the mutlisims maps would do
+    const auto nUniverses = m_signal_true_multisims.at(group).at(paramName).size();
+
+    // Get the nominal smearing matrix
+    const auto pSmearingNom = CrossSectionHelper::FlattenMatrix(this->GetSmearingMatrixNominal());
+    if (!pSmearingNom)
+        throw std::logic_error("CrossSection::GetSmearingMatrixDistributionParams - The nominal smearing matrix can't be calculated");
+
+    // Get the distirbution parameters
+    return this->GetDistributionParams([&](const unsigned int &iUni)
+    {
+        // This is the function we call to get the value in each universe
+        return CrossSectionHelper::FlattenMatrix(this->GetSmearingMatrixInUniverse(group, paramName, iUni));
+
+    }, nUniverses, pSmearingNom);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystBiasCovariancePair CrossSectionHelper::CrossSection::GetSmearingMatrixDistributionParamsUnisim(const std::string &group, const std::string &paramName, const std::string &cvName) const
+{
+    // Get the smearing matrix using the unisim sample, in the central value sample and in the nominal universe
+    const auto pSmearing = CrossSectionHelper::FlattenMatrix(this->GetSmearingMatrixForUnisim(group, paramName));
+    const auto pSmearingCV = CrossSectionHelper::FlattenMatrix(this->GetSmearingMatrixForUnisim(group, cvName));
+    const auto pSmearingNom = CrossSectionHelper::FlattenMatrix(this->GetSmearingMatrixNominal());
+
+    // Get the parameters
+    return this->GetDistributionParamsUnisim(pSmearing, pSmearingCV, pSmearingNom);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+std::shared_ptr<TH1F> CrossSectionHelper::GetTH1F(const std::vector<float> &binEdges)
+{
+    return std::make_shared<TH1F>(("xSecHist_" + std::to_string(m_histCount++)).c_str(), "", binEdges.size() - 1, binEdges.data());
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+std::shared_ptr<TH2F> CrossSectionHelper::GetTH2F(const std::vector<float> &binEdges)
+{
+    return std::make_shared<TH2F>(("xSecHist_" + std::to_string(m_histCount++)).c_str(), "", binEdges.size() - 1, binEdges.data(), binEdges.size() - 1, binEdges.data());
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystTH1FMap CrossSectionHelper::GetSystTH1FMap(const std::vector<float> &binEdges, const SystDimensionsMap &dimensions)
+{
+    SystTH1FMap map;
+
+    for (const auto &[paramName, nUniverses] : dimensions)
+    {
+        std::vector< std::shared_ptr<TH1F> > universes;
+        for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
+        {
+            universes.push_back(CrossSectionHelper::GetTH1F(binEdges));
+        }
+
+        map.emplace(paramName, universes);
+    }
+
+    return map;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystUnisimTH1FMap CrossSectionHelper::GetSystUnisimTH1FMap(const std::vector<float> &binEdges, const SystUnisimDimensionsMap &dimensions)
+{
+    SystUnisimTH1FMap map;
+
+    // Setup a vector to keep track of the used central-values sample names
+    std::vector<std::string> allCVNames;
+
+    for (const auto &[paramName, cvName] : dimensions)
+    {
+        // Include all parameter names
+        map.emplace(paramName, CrossSectionHelper::GetTH1F(binEdges));
+
+        // If we haven't seen this central value sample name before, then add it too!
+        if (std::find(allCVNames.begin(), allCVNames.end(), cvName) == allCVNames.end())
+        {
+            if (map.find(cvName) != map.end())
+                throw std::invalid_argument("CrossSectionHelper::GetSystUnisimTH1FMap - Name \"" + cvName + "\" is describes a parameter and a central-value sample");
+
+            map.emplace(cvName, CrossSectionHelper::GetTH1F(binEdges));
+        }
+    }
+
+    return map;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystTH2FMap CrossSectionHelper::GetSystTH2FMap(const std::vector<float> &binEdges, const SystDimensionsMap &dimensions)
+{
+    SystTH2FMap map;
+
+    for (const auto &[paramName, nUniverses] : dimensions)
+    {
+        std::vector< std::shared_ptr<TH2F> > universes;
+        for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
+        {
+            universes.push_back(CrossSectionHelper::GetTH2F(binEdges));
+        }
+
+        map.emplace(paramName, universes);
+    }
+
+    return map;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystUnisimTH2FMap CrossSectionHelper::GetSystUnisimTH2FMap(const std::vector<float> &binEdges, const SystUnisimDimensionsMap &dimensions)
+{
+    SystUnisimTH2FMap map;
+
+    // Setup a vector to keep track of the used central-values sample names
+    std::vector<std::string> allCVNames;
+
+    for (const auto &[paramName, cvName] : dimensions)
+    {
+        // Include all parameter names
+        map.emplace(paramName, CrossSectionHelper::GetTH2F(binEdges));
+
+        // If we haven't seen this central value sample name before, then add it too!
+        if (std::find(allCVNames.begin(), allCVNames.end(), cvName) == allCVNames.end())
+        {
+            if (map.find(cvName) != map.end())
+                throw std::invalid_argument("CrossSectionHelper::GetSystUnisimTH2FMap - Name \"" + cvName + "\" is describes a parameter and a central-value sample");
+
+            map.emplace(cvName, CrossSectionHelper::GetTH2F(binEdges));
+        }
+    }
+
+    return map;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystFloatMap CrossSectionHelper::GetWeightsMap(const Event::Truth &truth, const SystDimensionsMap &dimensions)
+{
+    if (!truth.systParamNames.IsSet() || !truth.systParamFirstValueIndex.IsSet() || !truth.systParamValues.IsSet())
+        throw std::invalid_argument("CrossSectionHelper::GetWeightsMap - Systematic parameters are not set in the input truth information");
+
+    const auto &systParamNames = truth.systParamNames();
+    const auto &systParamFirstValueIndex = truth.systParamFirstValueIndex();
+    const auto &systParamValues = truth.systParamValues();
+
+    // Get the total number of parameters available
+    const auto nParameters = systParamNames.size();
+
+    // Make a new map for the output
+    SystFloatMap map;
+
+    // Loop over all parameters in the input dimensions map
+    for (const auto &[paramName, nUniverses] : dimensions)
+    {
+        // Find the desired name
+        const auto iter = std::find(systParamNames.begin(), systParamNames.end(), paramName);
+        if (iter == systParamNames.end())
+            throw std::invalid_argument("CrossSectionHelper::GetWeightsMap - Unknown parameter: " + paramName);
+
+        // Get the index of the requested parameter
+        const unsigned int index = std::distance(systParamNames.begin(), iter);
+
+        // Get the first and last value in the weights vector
+        const unsigned int firstValueIndex = systParamFirstValueIndex.at(index);
+        const unsigned int lastValueIndex = ((index == nParameters - 1u) ? systParamValues.size() : systParamFirstValueIndex.at(index + 1u));
+
+        // Pick out the weights corresponding the the desired parameter
+        const std::vector<float> weights(std::next(systParamValues.begin(), firstValueIndex), std::next(systParamValues.begin(), lastValueIndex));
+
+        // Check that we have the right number of weights
+        if (weights.size() != nUniverses)
+            throw std::invalid_argument("CrossSectionHelper::GetWeightsMap - Number of weights for parameter " + paramName + " doesn't match the input dimensions");
+
+        // Setup the output map entry
+        auto &outputWeights = map[paramName];
+
+        // Fill the output weights
+        for (const auto &weight : weights)
+        {
+            // ATTN sometimes we have non-physical weights in the input. Here we decide what to do with them!
+
+            // If the weight is non-negative and non-infinite then it's okay!
+            if (weight >= 0.f && std::abs(std::abs(weight) - std::numeric_limits<float>::max()) > std::numeric_limits<float>::epsilon() && std::isfinite(weight))
+            {
+                outputWeights.push_back(weight);
+                continue;
+            }
+
+            // Force negative weights to be zero and infinite weights to be unity
+            outputWeights.push_back((weight < 0.f) ? 0.f : 1.f);
+        }
+    }
+
+    return map;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+CrossSectionHelper::SystFloatMap CrossSectionHelper::GetWeightsMap(const Event::Truth &truth, const SystDimensionsMap &dimensions, const SystMutuallyExclusiveDimensionsMap &mutuallyExclusiveDimensions)
+{
+    // Make a new map for the output
+    SystFloatMap map;
+
+    // Make new dimensions map to populate with the parameters that are available in the input truth information directly
+    SystDimensionsMap availableParamDimensions;
+
+    const auto &systParamNames = truth.systParamNames();
+
+    // Find which parameters (if any) are built from mutually exclusive parameters
+    for (const auto &[paramName, nUniverses] : dimensions)
+    {
+        // Check if this parameter name is available in the input truth information, if so just skip it for later
+        if (std::find(systParamNames.begin(), systParamNames.end(), paramName) != systParamNames.end())
+        {
+            availableParamDimensions.emplace(paramName, nUniverses);
+            continue;
+        }
+
+        // Insist that parameters not available in the input truth information are listed in the mutually exlusive dimensions
+        if (mutuallyExclusiveDimensions.find(paramName) == mutuallyExclusiveDimensions.end())
+            throw std::invalid_argument("CrossSectionHelper::GetWeightsMap - Parameter \"" + paramName + "\" isn't listed in the input truth information or in the supplied mutually exclusive parameter names");
+
+        // We have a mutually exclusive parameter, so get it's weights
+        const auto &[parameters, nUniversesCheck] = mutuallyExclusiveDimensions.at(paramName);
+        if (nUniverses != nUniversesCheck)
+            throw std::invalid_argument("CrossSectionHelper::GetWeightsMap - Inconsistent number of universes specified for parameter: \"" + paramName + "\"");
+
+        const auto weights = CrossSectionHelper::GetMutuallyExclusiveWeights(truth, parameters, nUniverses);
+
+        // Add the result to the output map
+        map.emplace(paramName, weights);
+    }
+
+    // Now get the weights from the rest of the parameters
+    auto availableParamWeights = CrossSectionHelper::GetWeightsMap(truth, availableParamDimensions);
+
+    // Combine this together and return it!
+    map.insert(availableParamWeights.begin(), availableParamWeights.end());
+    return map;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+std::vector<float> CrossSectionHelper::GetMutuallyExclusiveWeights(const Event::Truth &truth, const std::vector<std::string> &parameters, const unsigned int nUniverses)
+{
+    // Build a dimensions map using the same number of universes for each parameter
+    SystDimensionsMap dimensions;
+    for (const auto &paramName : parameters)
+        dimensions.emplace(paramName, nUniverses);
+
+    // Get the weights map for these parameters individually
+    const auto weightsMap = CrossSectionHelper::GetWeightsMap(truth, dimensions);
+
+    // Extract the mutually exclusive weights
+    std::vector<float> weights;
+    for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
+    {
+        // Find the parameter name for which the weight is not exactly one
+        float weight = 1.f;
+        bool foundNonUnitWeight = false;
+
+        for (const auto &paramName : parameters)
+        {
+            const auto universeWeight = weightsMap.at(paramName).at(iUni);
+            if (std::abs(universeWeight - 1.f) > std::numeric_limits<float>::epsilon())
+            {
+                // This weight is something other than one!
+
+                // If we find more than one weight that's not exactly one - then the parameters aren't mutually exclusive!
+                if (foundNonUnitWeight)
+                    throw std::logic_error("CrossSectionHelper::GetMutuallyExclusiveWeights - Input weights aren't mutually exclusive");
+
+                // Store this weight
+                weight = universeWeight;
+                foundNonUnitWeight = true;
+            }
+        }
+
+        // ATTN it's still posible that all universe weights are exactly one. This is okay - just use 1.f as the weight
+        weights.push_back(weight);
+    }
+
+    return weights;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+std::vector<float> CrossSectionHelper::GenerateBootstrapWeights(const unsigned int nUniverses)
+{
+    std::poisson_distribution<int> poisson(1.f);
+
+    std::vector<float> weights;
+    for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
+    {
+        weights.push_back(static_cast<float>(poisson(m_generator)));
+    }
+
+    return weights;
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+void CrossSectionHelper::FillSystTH1FMap(const float value, const float nominalWeight, const SystFloatMap &weights, SystTH1FMap &map)
+{
+    // Validate the dimensions of the input SystMaps
+    CrossSectionHelper::ValidateSystMap(weights, CrossSectionHelper::GetSystMapDimensions(map));
+
+    // Fill the histograms in each systematic universe
+    for (const auto &[paramName, universeWeights] : weights)
+    {
+        const auto nUniverses = universeWeights.size();
+        auto &histVector = map.at(paramName);
+
+        // ATTN this should never happen because of the above ValidateSystMap check, but including here for sanity
+        if (histVector.size() != nUniverses)
+            throw std::logic_error("CrossSectionHelper::FillSystTH1FMap - Input weights map and SystTH1FMap have different dimensions");
+
+        for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
+        {
+            // ATTN we use the product of the nominal weight and the universe weight
+            const auto weight = nominalWeight * universeWeights.at(iUni);
+
+            histVector.at(iUni)->Fill(value, weight);
+        }
+    }
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+void CrossSectionHelper::FillSystTH2FMap(const float xValue, const float yValue, const float nominalWeight, const SystFloatMap &weights, SystTH2FMap &map)
+{
+    // Validate the dimensions of the input SystMaps
+    CrossSectionHelper::ValidateSystMap(weights, CrossSectionHelper::GetSystMapDimensions(map));
+
+    // Fill the histograms in each systematic universe
+    for (const auto &[paramName, universeWeights] : weights)
+    {
+        const auto nUniverses = universeWeights.size();
+        auto &histVector = map.at(paramName);
+
+        // ATTN this should never happen because of the above ValidateSystMap check, but including here for sanity
+        if (histVector.size() != nUniverses)
+            throw std::logic_error("CrossSectionHelper::FillSystTH1FMap - Input weights map and SystTH1FMap have different dimensions");
+
+        for (unsigned int iUni = 0; iUni < nUniverses; ++iUni)
+        {
+            // ATTN we use the product of the nominal weight and the universe weight
+            const auto weight = nominalWeight * universeWeights.at(iUni);
+
+            histVector.at(iUni)->Fill(xValue, yValue, weight);
+        }
+    }
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::GetMatrixFromHist(const std::shared_ptr<TH1F> &pHist)
+{
+    std::vector<float> elements;
+
+    const unsigned int nBins = pHist->GetNbinsX();
+    for (unsigned int iBin = 1; iBin <= nBins; ++iBin)
+    {
+        elements.push_back(pHist->GetBinContent(iBin));
+    }
+
+    return ubsmear::UBMatrix(elements, nBins, 1);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+ubsmear::UBMatrix CrossSectionHelper::GetMatrixFromHist(const std::shared_ptr<TH2F> &pHist)
+{
+    std::vector< std::vector<float> > elements;
+
+    const unsigned int nBinsX = pHist->GetNbinsX();
+    const unsigned int nBinsY = pHist->GetNbinsY();
+
+    for (unsigned int iBinX = 1; iBinX <= nBinsX; ++iBinX)
+    {
+        // Make a new row
+        elements.emplace_back();
+        auto &row = elements.back();
+
+        for (unsigned int iBinY = 1; iBinY <= nBinsY; ++iBinY)
+        {
+            row.push_back(pHist->GetBinContent(iBinX, iBinY));
+        }
+    }
+
+    return ubsmear::UBMatrix(elements);
+}
+
+// -----------------------------------------------------------------------------------------------------------------------------------------
+
+std::shared_ptr<ubsmear::UBMatrix> CrossSectionHelper::FlattenMatrix(const std::shared_ptr<ubsmear::UBMatrix> &pMatrix)
+{
+    if (!pMatrix)
+        throw std::invalid_argument("CrossSectionHelper::FlattenMatrix - Input matrix pointer is null");
+
+    return std::make_shared<ubsmear::UBMatrix>(ubsmear::UBSmearingHelper::Flatten(*pMatrix));
 }
 
 } // namespace ubcc1pi
